Compute parameter estimate updates

Compute parameter estimate updates

Usage

est_params(
  hawkes,
  parameters,
  parent_est_mat,
  boundary = NULL,
  fixed_spatial = NULL,
  fixed_temporal = NULL
)

Arguments

hawkes

A hawkes object.

parameters

Named list containing background, triggering, spatial, and temporal parameters. To fix kernel parameters, include a fixed list such as fixed$spatial = c("mean").

parent_est_mat

Matrix returned by parent_est().

boundary

Optional boundary width used to correct edge bias.

fixed_spatial

Optional character vector naming spatial parameters to hold fixed.

fixed_temporal

Optional character vector naming temporal parameters to hold fixed.

Value

A named list mirroring parameters with updated estimates.

Examples

spatial_region <- create_rectangular_sf(0, 10, 0, 10)

params <- list(
  background_rate = list(intercept = -4),
  triggering_rate = 0.75,
  spatial = list(mean = 0, sd = 0.75),
  temporal = list(rate = 2)
)
hawkes <- rHawkes(params, time_window = c(0, 50), spatial_region = spatial_region)
parent_est_mat <- parent_est(hawkes, params)
est_params(hawkes, params, parent_est_mat)
#> $background_rate
#> $background_rate$intercept
#> [1] -3.87969
#> 
#> 
#> $triggering_rate
#> [1] 0.6624226
#> 
#> $spatial
#> $spatial$mean
#> [1] 0
#> 
#> $spatial$sd
#> [1] 0.7145118
#> 
#> 
#> $temporal
#> $temporal$rate
#> [1] 2.048687
#> 
#>