Estimate the covariance matrix of the MLE using observed information

Computes the observed information matrix (negative Hessian of the log-likelihood) and returns its inverse as an estimated covariance matrix for Hawkes model parameters.

Usage

hessian_est(hawkes, est)

Arguments

hawkes

A hawkes object.

est

A hawkes_fit object from hawkes_mle().

Value

A named covariance matrix corresponding to the estimated parameters.

Details

Requires the numDeriv package. The Hessian of the full log-likelihood is approximated with central differences. The negative Hessian is inverted and its row and column names are matched to the flattened parameter vector.

Examples

spatial_region <- create_rectangular_sf(0, 10, 0, 10)

params <- list(
  background_rate = list(intercept = -4),
  triggering_rate = 0.75,
  spatial = list(mean = 0, sd = 0.75),
  temporal = list(rate = 2)
)
hawkes <- rHawkes(params, time_window = c(0, 50), spatial_region = spatial_region)
est <- hawkes_mle(hawkes, inits = params)
hessian_est(hawkes, est$est)
#>                           background_rate.intercept triggering_rate
#> background_rate.intercept              0.0177373325   -0.0029001104
#> triggering_rate                       -0.0029001104    0.0035683583
#> spatial.mean                          -0.0005931944    0.0002289469
#> spatial.sd                            -0.0014973329    0.0005710269
#> temporal.rate                          0.0082045070   -0.0033010719
#>                            spatial.mean    spatial.sd temporal.rate
#> background_rate.intercept -0.0005931944 -0.0014973329  0.0082045070
#> triggering_rate            0.0002289469  0.0005710269 -0.0033010719
#> spatial.mean               0.0031028638  0.0005055939 -0.0008694766
#> spatial.sd                 0.0005055939  0.0018952644  0.0005997227
#> temporal.rate             -0.0008694766  0.0005997227  0.0666724667