Generate a Hawkes process

Generate a Hawkes process

Usage

rHawkes(
  params,
  time_window,
  spatial_region,
  covariate_columns = NULL,
  temporal_burnin = (time_window[2] - time_window[1])/(10),
  spatial_burnin = sum(as.numeric(sf::st_area(spatial_region)))^0.25,
  temporal_family = "Exponential",
  spatial_family = "Gaussian"
)

Arguments

params

Named list containing background, triggering, spatial, and temporal parameters. See the examples for the expected structure.

time_window

Numeric vector of length two specifying the simulated window.

spatial_region

sf object defining the spatial region.

covariate_columns

Optional character vector naming background covariates.

temporal_burnin

Temporal burn-in duration. Defaults to one tenth of the window length.

spatial_burnin

Spatial burn-in radius. Defaults to area(spatial_region)^0.25.

temporal_family

Temporal triggering kernel. Defaults to "Exponential". Other options include "Power Law", "Uniform", and "Gamma".

spatial_family

Spatial triggering kernel. Defaults to "Gaussian". Other options include "Uniform" and "Exponential".

Value

A hawkes object with a generated Hawkes process.

Examples

spatial_region <- create_rectangular_sf(0, 10, 0, 10)

params <- list(
  background_rate = list(intercept = -4),
  triggering_rate = 0.75,
  spatial = list(mean = 0, sd = 0.75),
  temporal = list(rate = 2)
)
(hawkes <- rHawkes(
  params,
  time_window = c(0, 50),
  spatial_region = spatial_region,
  spatial_burnin = 1
))
#> <hawkes object>
#> 
#> Number of events: 259 
#> 
#> Spatial Region:
#> Simple feature collection with 1 feature and 2 fields
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 0 ymin: 0 xmax: 10 ymax: 10
#> CRS:           NA
#>                         geometry geoid area
#> 1 POLYGON ((0 0, 10 0, 10 10,...     1  100
#> 
#> Time Window:
#> [1]  0 50
#> 
#> Event data (first 10 rows):
#>             x        y         t  id parent gen family
#> 135 3.3598916 8.670063 0.1361642  18    129   1    129
#> 12  1.0471998 9.085512 0.1527087 109      0   0    109
#> 136 1.0206951 9.310438 0.3050455  19    109   1    109
#> 14  7.9780756 9.006029 0.3359881  62      0   0     62
#> 137 9.0113325 9.628210 0.4085357  20     62   1     62
#> 203 0.7074479 8.605957 0.4950205 115     19   2    109
#> 138 8.2426954 8.752003 0.5122444  21     62   1     62
#> 202 2.6957914 8.145115 0.6876934 113     16   2     41
#> 204 0.5167700 9.685695 0.6989371 116     19   2    109
#> 249 0.7666737 8.891519 1.2260909 167    116   3    109
#>                       geometry
#> 135  POINT (3.359892 8.670063)
#> 12     POINT (1.0472 9.085512)
#> 136  POINT (1.020695 9.310438)
#> 14   POINT (7.978076 9.006029)
#> 137   POINT (9.011332 9.62821)
#> 203 POINT (0.7074479 8.605957)
#> 138  POINT (8.242695 8.752003)
#> 202  POINT (2.695791 8.145115)
#> 204   POINT (0.51677 9.685695)
#> 249 POINT (0.7666737 8.891519)
#> # 249 more rows
#> # Use `print(n = ...)` to see more rows

params <- list(
  background_rate = list(intercept = -5, z = 1),
  triggering_rate = 0.5,
  spatial = list(mean = 0, sd = 0.75),
  temporal = list(rate = 2),
  fixed = list(spatial = "mean")
)
data("example_background_covariates")
rHawkes(
  params,
  c(0, 50),
  example_background_covariates,
  covariate_columns = c("z"),
  spatial_burnin = 1
)
#> <hawkes object>
#> 
#> Number of events: 137 
#> 
#> Spatial Region:
#> Simple feature collection with 121 features and 3 fields
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 0 ymin: 0 xmax: 10 ymax: 10
#> CRS:           NA
#> First 10 features:
#>    geoid      area          z                       geometry
#> 1      1 0.8264463 -0.5495721 POLYGON ((0 0, 0.9090909 0,...
#> 2      2 0.8264463 -0.5491228 POLYGON ((0.9090909 0, 1.81...
#> 3      3 0.8264463 -0.5471802 POLYGON ((1.818182 0, 2.727...
#> 4      4 0.8264463 -0.5421628 POLYGON ((2.727273 0, 3.636...
#> 5      5 0.8264463 -0.5350741 POLYGON ((3.636364 0, 4.545...
#> 6      6 0.8264463 -0.5314488 POLYGON ((4.545455 0, 5.454...
#> 7      7 0.8264463 -0.5350741 POLYGON ((5.454545 0, 6.363...
#> 8      8 0.8264463 -0.5421628 POLYGON ((6.363636 0, 7.272...
#> 9      9 0.8264463 -0.5471802 POLYGON ((7.272727 0, 8.181...
#> 10    10 0.8264463 -0.5491228 POLYGON ((8.181818 0, 9.090...
#> 
#> Time Window:
#> [1]  0 50
#> 
#> Event data (first 10 rows):
#>           x        y        t id parent gen family          z      area
#> 1  6.031994 6.233228 1.453387 66      0   0     66  2.4678587 0.8264463
#> 2  9.632759 8.662491 2.234126 72      0   0     72 -0.5491228 0.8264463
#> 3  2.306170 1.365860 2.352277  3      0   0      3 -0.5314488 0.8264463
#> 4  4.015243 4.668486 2.906517 35      0   0     35  3.2187604 0.8264463
#> 5  3.264748 3.323610 3.787373 56     35   1     35  0.2457622 0.8264463
#> 6  5.520873 3.926571 4.086836 21      0   0     21  2.4678587 0.8264463
#> 7  4.132392 5.157979 6.673819 29      0   0     29  3.2187604 0.8264463
#> 8  6.012174 4.751413 8.061339 54      0   0     54  3.2187604 0.8264463
#> 9  5.669745 5.583916 8.228534 57     54   1     54  2.4678587 0.8264463
#> 10 5.670650 5.512628 8.563846 94     57   2     54  2.4678587 0.8264463
#>                     geometry
#> 1  POINT (6.031994 6.233228)
#> 2  POINT (9.632759 8.662491)
#> 3    POINT (2.30617 1.36586)
#> 4  POINT (4.015243 4.668486)
#> 5   POINT (3.264748 3.32361)
#> 6  POINT (5.520873 3.926571)
#> 7  POINT (4.132392 5.157979)
#> 8  POINT (6.012174 4.751413)
#> 9  POINT (5.669745 5.583916)
#> 10  POINT (5.67065 5.512628)
#> # 127 more rows
#> # Use `print(n = ...)` to see more rows