Compute the conditional intensity at a specified time

Compute the conditional intensity at a specified time

Usage

spatial_conditional_intensity(hawkes, parameters, time, stepsize)

Arguments

hawkes

A hawkes object.

parameters

Named list containing background, triggering, spatial, and temporal parameters.

time

Time point where the conditional intensity is evaluated.

stepsize

Grid cell size used to evaluate the spatial intensity.

Value

A numeric vector.

Examples

spatial_region <- create_rectangular_sf(0, 10, 0, 10)

params <- list(
  background_rate = list(intercept = -4),
  triggering_rate = 0.75,
  spatial = list(mean = 0, sd = 0.75),
  temporal = list(rate = 2)
)
hawkes <- rHawkes(params, time_window = c(0, 50), spatial_region = spatial_region)
spatial_conditional_intensity(hawkes, params, 25, 0.5)
#>        x    y  t  intensity
#> 1   0.25 0.25 25 0.01854505
#> 2   0.75 0.25 25 0.01918719
#> 3   1.25 0.25 25 0.02048178
#> 4   1.75 0.25 25 0.02206567
#> 5   2.25 0.25 25 0.02378161
#> 6   2.75 0.25 25 0.02680231
#> 7   3.25 0.25 25 0.03105533
#> 8   3.75 0.25 25 0.03269138
#> 9   4.25 0.25 25 0.02916420
#> 10  4.75 0.25 25 0.02362030
#> 11  5.25 0.25 25 0.02001482
#> 12  5.75 0.25 25 0.01895832
#> 13  6.25 0.25 25 0.02021497
#> 14  6.75 0.25 25 0.02563372
#> 15  7.25 0.25 25 0.03698261
#> 16  7.75 0.25 25 0.04904598
#> 17  8.25 0.25 25 0.05090938
#> 18  8.75 0.25 25 0.04069891
#> 19  9.25 0.25 25 0.02860198
#> 20  9.75 0.25 25 0.02199192
#> 21  0.25 0.75 25 0.01879554
#> 22  0.75 0.75 25 0.02013426
#> 23  1.25 0.75 25 0.02278926
#> 24  1.75 0.75 25 0.02575581
#> 25  2.25 0.75 25 0.02796352
#> 26  2.75 0.75 25 0.03099884
#> 27  3.25 0.75 25 0.03570480
#> 28  3.75 0.75 25 0.03751617
#> 29  4.25 0.75 25 0.03281666
#> 30  4.75 0.75 25 0.02544936
#> 31  5.25 0.75 25 0.02059829
#> 32  5.75 0.75 25 0.01899527
#> 33  6.25 0.75 25 0.01967541
#> 34  6.75 0.75 25 0.02335734
#> 35  7.25 0.75 25 0.03103867
#> 36  7.75 0.75 25 0.03915180
#> 37  8.25 0.75 25 0.04046048
#> 38  8.75 0.75 25 0.03396684
#> 39  9.25 0.75 25 0.02664004
#> 40  9.75 0.75 25 0.02296919
#> 41  0.25 1.25 25 0.01895970
#> 42  0.75 1.25 25 0.02075468
#> 43  1.25 1.25 25 0.02429956
#> 44  1.75 1.25 25 0.02816978
#> 45  2.25 1.25 25 0.03072948
#> 46  2.75 1.25 25 0.03395638
#> 47  3.25 1.25 25 0.03939253
#> 48  3.75 1.25 25 0.04175626
#> 49  4.25 1.25 25 0.03628424
#> 50  4.75 1.25 25 0.02730160
#> 51  5.25 1.25 25 0.02122161
#> 52  5.75 1.25 25 0.01902139
#> 53  6.25 1.25 25 0.01900069
#> 54  6.75 1.25 25 0.02059922
#> 55  7.25 1.25 25 0.02396112
#> 56  7.75 1.25 25 0.02749120
#> 57  8.25 1.25 25 0.02823481
#> 58  8.75 1.25 25 0.02618472
#> 59  9.25 1.25 25 0.02455437
#> 60  9.75 1.25 25 0.02442407
#> 61  0.25 1.75 25 0.01887016
#> 62  0.75 1.75 25 0.02041802
#> 63  1.25 1.75 25 0.02350237
#> 64  1.75 1.75 25 0.02707370
#> 65  2.25 1.75 25 0.03034858
#> 66  2.75 1.75 25 0.03598355
#> 67  3.25 1.75 25 0.04485114
#> 68  3.75 1.75 25 0.04921431
#> 69  4.25 1.75 25 0.04255018
#> 70  4.75 1.75 25 0.03064079
#> 71  5.25 1.75 25 0.02235000
#> 72  5.75 1.75 25 0.01919796
#> 73  6.25 1.75 25 0.01862315
#> 74  6.75 1.75 25 0.01901705
#> 75  7.25 1.75 25 0.01997576
#> 76  7.75 1.75 25 0.02099982
#> 77  8.25 1.75 25 0.02146035
#> 78  8.75 1.75 25 0.02182958
#> 79  9.25 1.75 25 0.02323530
#> 80  9.75 1.75 25 0.02495474
#> 81  0.25 2.25 25 0.01862208
#> 82  0.75 2.25 25 0.01948263
#> 83  1.25 2.25 25 0.02125298
#> 84  1.75 2.25 25 0.02369631
#> 85  2.25 2.25 25 0.02754550
#> 86  2.75 2.25 25 0.03595018
#> 87  3.25 2.25 25 0.04837496
#> 88  3.75 2.25 25 0.05470193
#> 89  4.25 2.25 25 0.04723974
#> 90  4.75 2.25 25 0.03313564
#> 91  5.25 2.25 25 0.02319160
#> 92  5.75 2.25 25 0.01935354
#> 93  6.25 2.25 25 0.01849956
#> 94  6.75 2.25 25 0.01847493
#> 95  7.25 2.25 25 0.01865210
#> 96  7.75 2.25 25 0.01887282
#> 97  8.25 2.25 25 0.01916543
#> 98  8.75 2.25 25 0.01996817
#> 99  9.25 2.25 25 0.02166114
#> 100 9.75 2.25 25 0.02332813
#> 101 0.25 2.75 25 0.01842461
#> 102 0.75 2.75 25 0.01873604
#> 103 1.25 2.75 25 0.01943059
#> 104 1.75 2.75 25 0.02074713
#> 105 2.25 2.75 25 0.02402416
#> 106 2.75 2.75 25 0.03191368
#> 107 3.25 2.75 25 0.04325184
#> 108 3.75 2.75 25 0.04905459
#> 109 4.25 2.75 25 0.04287425
#> 110 4.75 2.75 25 0.03092861
#> 111 5.25 2.75 25 0.02247542
#> 112 5.75 2.75 25 0.01920236
#> 113 6.25 2.75 25 0.01844921
#> 114 6.75 2.75 25 0.01835352
#> 115 7.25 2.75 25 0.01836824
#> 116 7.75 2.75 25 0.01841165
#> 117 8.25 2.75 25 0.01855261
#> 118 8.75 2.75 25 0.01901346
#> 119 9.25 2.75 25 0.01992548
#> 120 9.75 2.75 25 0.02079324
#> 121 0.25 3.25 25 0.01834104
#> 122 0.75 3.25 25 0.01841798
#> 123 1.25 3.25 25 0.01862458
#> 124 1.75 3.25 25 0.01920755
#> 125 2.25 3.25 25 0.02108041
#> 126 2.75 3.25 25 0.02569051
#> 127 3.25 3.25 25 0.03219736
#> 128 3.75 3.25 25 0.03550765
#> 129 4.25 3.25 25 0.03207243
#> 130 4.75 3.25 25 0.02541632
#> 131 5.25 3.25 25 0.02072913
#> 132 5.75 3.25 25 0.01893412
#> 133 6.25 3.25 25 0.01850575
#> 134 6.75 3.25 25 0.01840223
#> 135 7.25 3.25 25 0.01835509
#> 136 7.75 3.25 25 0.01834109
#> 137 8.25 3.25 25 0.01837882
#> 138 8.75 3.25 25 0.01852926
#> 139 9.25 3.25 25 0.01882544
#> 140 9.75 3.25 25 0.01910536
#> 141 0.25 3.75 25 0.01832022
#> 142 0.75 3.75 25 0.01833759
#> 143 1.25 3.75 25 0.01840263
#> 144 1.75 3.75 25 0.01863639
#> 145 2.25 3.75 25 0.01938747
#> 146 2.75 3.75 25 0.02112872
#> 147 3.25 3.75 25 0.02349954
#> 148 3.75 3.75 25 0.02468901
#> 149 4.25 3.75 25 0.02350429
#> 150 4.75 3.75 25 0.02134155
#> 151 5.25 3.75 25 0.02012931
#> 152 5.75 3.75 25 0.01992087
#> 153 6.25 3.75 25 0.01972759
#> 154 6.75 3.75 25 0.01920371
#> 155 7.25 3.75 25 0.01869276
#> 156 7.75 3.75 25 0.01842786
#> 157 8.25 3.75 25 0.01835054
#> 158 8.75 3.75 25 0.01836282
#> 159 9.25 3.75 25 0.01842073
#> 160 9.75 3.75 25 0.01847773
#> 161 0.25 4.25 25 0.01831720
#> 162 0.75 4.25 25 0.01832487
#> 163 1.25 4.25 25 0.01835947
#> 164 1.75 4.25 25 0.01847239
#> 165 2.25 4.25 25 0.01874781
#> 166 2.75 4.25 25 0.01925420
#> 167 3.25 4.25 25 0.01988731
#> 168 3.75 4.25 25 0.02032429
#> 169 4.25 4.25 25 0.02066771
#> 170 4.75 4.25 25 0.02210835
#> 171 5.25 4.25 25 0.02549449
#> 172 5.75 4.25 25 0.02874865
#> 173 6.25 4.25 25 0.02845370
#> 174 6.75 4.25 25 0.02479215
#> 175 7.25 4.25 25 0.02107537
#> 176 7.75 4.25 25 0.01912652
#> 177 8.25 4.25 25 0.01849215
#> 178 8.75 4.25 25 0.01835647
#> 179 9.25 4.25 25 0.01834242
#> 180 9.75 4.25 25 0.01834478
#> 181 0.25 4.75 25 0.01831738
#> 182 0.75 4.75 25 0.01832277
#> 183 1.25 4.75 25 0.01834665
#> 184 1.75 4.75 25 0.01842203
#> 185 2.25 4.75 25 0.01859330
#> 186 2.75 4.75 25 0.01891640
#> 187 3.25 4.75 25 0.01948808
#> 188 3.75 4.75 25 0.02060848
#> 189 4.25 4.75 25 0.02385449
#> 190 4.75 4.75 25 0.03339217
#> 191 5.25 4.75 25 0.05122818
#> 192 5.75 4.75 25 0.06724707
#> 193 6.25 4.75 25 0.06600956
#> 194 6.75 4.75 25 0.04882461
#> 195 7.25 4.75 25 0.03136391
#> 196 7.75 4.75 25 0.02219939
#> 197 8.25 4.75 25 0.01921729
#> 198 8.75 4.75 25 0.01858221
#> 199 9.25 4.75 25 0.01848799
#> 200 9.75 4.75 25 0.01843748
#> 201 0.25 5.25 25 0.01831851
#> 202 0.75 5.25 25 0.01832239
#> 203 1.25 5.25 25 0.01833937
#> 204 1.75 5.25 25 0.01841695
#> 205 2.25 5.25 25 0.01872895
#> 206 2.75 5.25 25 0.01977372
#> 207 3.25 5.25 25 0.02237072
#> 208 3.75 5.25 25 0.02738751
#> 209 4.25 5.25 25 0.03842273
#> 210 4.75 5.25 25 0.06704725
#> 211 5.25 5.25 25 0.11988187
#> 212 5.75 5.25 25 0.16741054
#> 213 6.25 5.25 25 0.16323140
#> 214 6.75 5.25 25 0.11112390
#> 215 7.25 5.25 25 0.05831443
#> 216 7.75 5.25 25 0.03066651
#> 217 8.25 5.25 25 0.02182580
#> 218 8.75 5.25 25 0.02015994
#> 219 9.25 5.25 25 0.01995072
#> 220 9.75 5.25 25 0.01951641
#> 221 0.25 5.75 25 0.01832055
#> 222 0.75 5.75 25 0.01832453
#> 223 1.25 5.75 25 0.01834488
#> 224 1.75 5.75 25 0.01851204
#> 225 2.25 5.75 25 0.01953556
#> 226 2.75 5.75 25 0.02373375
#> 227 3.25 5.75 25 0.03467850
#> 228 3.75 5.75 25 0.05300341
#> 229 4.25 5.75 25 0.07923280
#> 230 4.75 5.75 25 0.13137581
#> 231 5.25 5.75 25 0.22666923
#> 232 5.75 5.75 25 0.31389464
#> 233 6.25 5.75 25 0.30367965
#> 234 6.75 5.75 25 0.20179386
#> 235 7.25 5.75 25 0.09936307
#> 236 7.75 5.75 25 0.04646680
#> 237 8.25 5.75 25 0.03062065
#> 238 8.75 5.75 25 0.02889976
#> 239 9.25 5.75 25 0.02880610
#> 240 9.75 5.75 25 0.02606859
#> 241 0.25 6.25 25 0.01833383
#> 242 0.75 6.25 25 0.01834626
#> 243 1.25 6.25 25 0.01838716
#> 244 1.75 6.25 25 0.01879952
#> 245 2.25 6.25 25 0.02164404
#> 246 2.75 6.25 25 0.03385968
#> 247 3.25 6.25 25 0.06579816
#> 248 3.75 6.25 25 0.11488533
#> 249 4.25 6.25 25 0.16186495
#> 250 4.75 6.25 25 0.21557210
#> 251 5.25 6.25 25 0.31135370
#> 252 5.75 6.25 25 0.40465883
#> 253 6.25 6.25 25 0.38645258
#> 254 6.75 6.25 25 0.25908919
#> 255 7.25 6.25 25 0.13450821
#> 256 7.75 6.25 25 0.07373320
#> 257 8.25 6.25 25 0.05947997
#> 258 8.75 6.25 25 0.06229922
#> 259 9.25 6.25 25 0.06228536
#> 260 9.75 6.25 25 0.05051917
#> 261 0.25 6.75 25 0.01846904
#> 262 0.75 6.75 25 0.01856174
#> 263 1.25 6.75 25 0.01863612
#> 264 1.75 6.75 25 0.01932947
#> 265 2.25 6.75 25 0.02465327
#> 266 2.75 6.75 25 0.04805335
#> 267 3.25 6.75 25 0.10920992
#> 268 3.75 6.75 25 0.19944317
#> 269 4.25 6.75 25 0.26459317
#> 270 4.75 6.75 25 0.28521095
#> 271 5.25 6.75 25 0.31672747
#> 272 5.75 6.75 25 0.36187692
#> 273 6.25 6.75 25 0.33943103
#> 274 6.75 6.75 25 0.24350929
#> 275 7.25 6.75 25 0.16019421
#> 276 7.75 6.75 25 0.13041131
#> 277 8.25 6.75 25 0.13218970
#> 278 8.75 6.75 25 0.14212710
#> 279 9.25 6.75 25 0.13783996
#> 280 9.75 6.75 25 0.10433017
#> 281 0.25 7.25 25 0.01935431
#> 282 0.75 7.25 25 0.01996056
#> 283 1.25 7.25 25 0.02006737
#> 284 1.75 7.25 25 0.02044038
#> 285 2.25 7.25 25 0.02650457
#> 286 2.75 7.25 25 0.05524388
#> 287 3.25 7.25 25 0.13088119
#> 288 3.75 7.25 25 0.24065839
#> 289 4.25 7.25 25 0.30864543
#> 290 4.75 7.25 25 0.29194581
#> 291 5.25 7.25 25 0.25287759
#> 292 5.75 7.25 25 0.24152737
#> 293 6.25 7.25 25 0.22695634
#> 294 6.75 7.25 25 0.20154142
#> 295 7.25 7.25 25 0.20375938
#> 296 7.75 7.25 25 0.23009141
#> 297 8.25 7.25 25 0.24721233
#> 298 8.75 7.25 25 0.25038737
#> 299 9.25 7.25 25 0.22926506
#> 300 9.75 7.25 25 0.16620467
#> 301 0.25 7.75 25 0.02290480
#> 302 0.75 7.75 25 0.02556428
#> 303 1.25 7.75 25 0.02572127
#> 304 1.75 7.75 25 0.02390637
#> 305 2.25 7.75 25 0.02645536
#> 306 2.75 7.75 25 0.04816755
#> 307 3.25 7.75 25 0.10808936
#> 308 3.75 7.75 25 0.19512822
#> 309 4.25 7.75 25 0.24623457
#> 310 4.75 7.75 25 0.22155220
#> 311 5.25 7.75 25 0.16857345
#> 312 5.75 7.75 25 0.14324116
#> 313 6.25 7.75 25 0.14942516
#> 314 6.75 7.75 25 0.19005407
#> 315 7.25 7.75 25 0.26837428
#> 316 7.75 7.75 25 0.33496339
#> 317 8.25 7.75 25 0.34206775
#> 318 8.75 7.75 25 0.31135624
#> 319 9.25 7.75 25 0.26216591
#> 320 9.75 7.75 25 0.18255137
#> 321 0.25 8.25 25 0.03133642
#> 322 0.75 8.25 25 0.03887147
#> 323 1.25 8.25 25 0.03915588
#> 324 1.75 8.25 25 0.03224093
#> 325 2.25 8.25 25 0.02708226
#> 326 2.75 8.25 25 0.03482545
#> 327 3.25 8.25 25 0.06460414
#> 328 3.75 8.25 25 0.10968517
#> 329 4.25 8.25 25 0.13800389
#> 330 4.75 8.25 25 0.12922109
#> 331 5.25 8.25 25 0.10666984
#> 332 5.75 8.25 25 0.10158536
#> 333 6.25 8.25 25 0.12868629
#> 334 6.75 8.25 25 0.20358702
#> 335 7.25 8.25 25 0.31254025
#> 336 7.75 8.25 25 0.37927347
#> 337 8.25 8.25 25 0.35286941
#> 338 8.75 8.25 25 0.27914810
#> 339 9.25 8.25 25 0.20795799
#> 340 9.75 8.25 25 0.13738406
#> 341 0.25 8.75 25 0.04200824
#> 342 0.75 8.75 25 0.05571513
#> 343 1.25 8.75 25 0.05617942
#> 344 1.75 8.75 25 0.04301868
#> 345 2.25 8.75 25 0.02957190
#> 346 2.75 8.75 25 0.02600605
#> 347 3.25 8.75 25 0.03423098
#> 348 3.75 8.75 25 0.05028105
#> 349 4.25 8.75 25 0.06479807
#> 350 4.75 8.75 25 0.07297929
#> 351 5.25 8.75 25 0.08068725
#> 352 5.75 8.75 25 0.09505174
#> 353 6.25 8.75 25 0.12860357
#> 354 6.75 8.75 25 0.20227405
#> 355 7.25 8.75 25 0.29895053
#> 356 7.75 8.75 25 0.34242632
#> 357 8.25 8.75 25 0.29106704
#> 358 8.75 8.75 25 0.19838721
#> 359 9.25 8.75 25 0.12516019
#> 360 9.75 8.75 25 0.07668967
#> 361 0.25 9.25 25 0.04595877
#> 362 0.75 9.25 25 0.06195024
#> 363 1.25 9.25 25 0.06248060
#> 364 1.75 9.25 25 0.04700385
#> 365 2.25 9.25 25 0.03046207
#> 366 2.75 9.25 25 0.02259466
#> 367 3.25 9.25 25 0.02255958
#> 368 3.75 9.25 25 0.02755432
#> 369 4.25 9.25 25 0.03707848
#> 370 4.75 9.25 25 0.05210469
#> 371 5.25 9.25 25 0.07068898
#> 372 5.75 9.25 25 0.08830013
#> 373 6.25 9.25 25 0.11222201
#> 374 6.75 9.25 25 0.16167610
#> 375 7.25 9.25 25 0.22611524
#> 376 7.75 9.25 25 0.24985184
#> 377 8.25 9.25 25 0.20291798
#> 378 8.75 9.25 25 0.12589607
#> 379 9.25 9.25 25 0.06883757
#> 380 9.75 9.25 25 0.03977992
#> 381 0.25 9.75 25 0.03899546
#> 382 0.75 9.75 25 0.05095852
#> 383 1.25 9.75 25 0.05135372
#> 384 1.75 9.75 25 0.03975920
#> 385 2.25 9.75 25 0.02726986
#> 386 2.75 9.75 25 0.02089049
#> 387 3.25 9.75 25 0.01961035
#> 388 3.75 9.75 25 0.02169982
#> 389 4.25 9.75 25 0.02836140
#> 390 4.75 9.75 25 0.04098231
#> 391 5.25 9.75 25 0.05585117
#> 392 5.75 9.75 25 0.06651362
#> 393 6.25 9.75 25 0.07636207
#> 394 6.75 9.75 25 0.09885192
#> 395 7.25 9.75 25 0.13141930
#> 396 7.75 9.75 25 0.14428347
#> 397 8.25 9.75 25 0.11844467
#> 398 8.75 9.75 25 0.07408573
#> 399 9.25 9.75 25 0.04077845
#> 400 9.75 9.75 25 0.02546742

params <- list(
  background_rate = list(intercept = -4.5, z = 1),
  triggering_rate = 0.5,
  spatial = list(mean = 0, sd = 0.75),
  temporal = list(rate = 2),
  fixed = list(spatial = "mean")
)
data("example_background_covariates")
hawkes <- rHawkes(
  params,
  c(0, 50),
  example_background_covariates,
  covariate_columns = "z",
  spatial_burnin = 1
)

spatial_conditional_intensity(hawkes, params, 25, 0.5)
#>        x    y  t   intensity
#> 1   0.25 0.25 25 0.006412589
#> 2   0.75 0.25 25 0.006412478
#> 3   0.25 0.75 25 0.006412506
#> 4   0.75 0.75 25 0.006412413
#> 5   1.25 0.25 25 0.006415160
#> 6   1.75 0.25 25 0.006415023
#> 7   1.25 0.75 25 0.006415127
#> 8   1.75 0.75 25 0.006415013
#> 9   2.25 0.25 25 0.006427445
#> 10  2.25 0.75 25 0.006427443
#> 11  2.75 0.25 25 0.006459765
#> 12  3.25 0.25 25 0.006459775
#> 13  2.75 0.75 25 0.006459768
#> 14  3.25 0.75 25 0.006459811
#> 15  3.75 0.25 25 0.006505816
#> 16  4.25 0.25 25 0.006506520
#> 17  3.75 0.75 25 0.006506047
#> 18  4.25 0.75 25 0.006507750
#> 19  4.75 0.25 25 0.006537580
#> 20  5.25 0.25 25 0.006604126
#> 21  4.75 0.75 25 0.006545218
#> 22  5.25 0.75 25 0.006660189
#> 23  5.75 0.25 25 0.006974901
#> 24  6.25 0.25 25 0.008418925
#> 25  5.75 0.75 25 0.007309443
#> 26  6.25 0.75 25 0.009769103
#> 27  6.75 0.25 25 0.011473747
#> 28  7.25 0.25 25 0.014888494
#> 29  6.75 0.75 25 0.015004502
#> 30  7.25 0.75 25 0.020820956
#> 31  7.75 0.25 25 0.015513030
#> 32  7.75 0.75 25 0.021907071
#> 33  8.25 0.25 25 0.012694543
#> 34  8.75 0.25 25 0.009197808
#> 35  8.25 0.75 25 0.017113705
#> 36  8.75 0.75 25 0.011156180
#> 37  9.25 0.25 25 0.007202808
#> 38  9.75 0.25 25 0.006556138
#> 39  9.25 0.75 25 0.007759267
#> 40  9.75 0.75 25 0.006657518
#> 41  0.25 1.25 25 0.006415189
#> 42  0.75 1.25 25 0.006415139
#> 43  0.25 1.75 25 0.006415038
#> 44  0.75 1.75 25 0.006415021
#> 45  1.25 1.25 25 0.006436385
#> 46  1.75 1.25 25 0.006436324
#> 47  1.25 1.75 25 0.006436327
#> 48  1.75 1.75 25 0.006436313
#> 49  2.25 1.25 25 0.006529354
#> 50  2.25 1.75 25 0.006529413
#> 51  2.75 1.25 25 0.006775987
#> 52  3.25 1.25 25 0.006776164
#> 53  2.75 1.75 25 0.006776435
#> 54  3.25 1.75 25 0.006778399
#> 55  3.75 1.25 25 0.007141422
#> 56  4.25 1.25 25 0.007145231
#> 57  3.75 1.75 25 0.007149152
#> 58  4.25 1.75 25 0.007165132
#> 59  4.75 1.25 25 0.007359135
#> 60  5.25 1.25 25 0.007493153
#> 61  4.75 1.75 25 0.007399525
#> 62  5.25 1.75 25 0.007538115
#> 63  5.75 1.25 25 0.008042339
#> 64  6.25 1.25 25 0.010730002
#> 65  5.75 1.75 25 0.007912528
#> 66  6.25 1.75 25 0.009805325
#> 67  6.75 1.25 25 0.016127345
#> 68  7.25 1.25 25 0.022471923
#> 69  6.75 1.75 25 0.013434311
#> 70  7.25 1.75 25 0.017820953
#> 71  7.75 1.25 25 0.023441600
#> 72  7.75 1.75 25 0.018390828
#> 73  8.25 1.25 25 0.018124025
#> 74  8.75 1.25 25 0.011615646
#> 75  8.25 1.75 25 0.014625816
#> 76  8.75 1.75 25 0.010064450
#> 77  9.25 1.25 25 0.007886629
#> 78  9.75 1.25 25 0.006683078
#> 79  9.25 1.75 25 0.007445788
#> 80  9.75 1.75 25 0.006602758
#> 81  0.25 2.25 25 0.006427450
#> 82  0.75 2.25 25 0.006427448
#> 83  1.25 2.25 25 0.006529365
#> 84  1.75 2.25 25 0.006529447
#> 85  2.25 2.25 25 0.006989718
#> 86  2.75 2.25 25 0.008337418
#> 87  3.25 2.25 25 0.008363551
#> 88  3.75 2.25 25 0.010784799
#> 89  4.25 2.25 25 0.010922086
#> 90  4.75 2.25 25 0.012549721
#> 91  5.25 2.25 25 0.012791821
#> 92  5.75 2.25 25 0.011755444
#> 93  6.25 2.25 25 0.012600839
#> 94  6.75 2.25 25 0.011833026
#> 95  7.25 2.25 25 0.013535116
#> 96  7.75 2.25 25 0.012393762
#> 97  8.25 2.25 25 0.010224101
#> 98  8.75 2.25 25 0.008161250
#> 99  9.25 2.25 25 0.006890636
#> 100 9.75 2.25 25 0.006511791
#> 101 0.25 2.75 25 0.006459767
#> 102 0.75 2.75 25 0.006459779
#> 103 0.25 3.25 25 0.006459770
#> 104 0.75 3.25 25 0.006459821
#> 105 1.25 2.75 25 0.006776050
#> 106 1.75 2.75 25 0.006776787
#> 107 1.25 3.25 25 0.006776386
#> 108 1.75 3.25 25 0.006780587
#> 109 2.25 2.75 25 0.008339078
#> 110 2.25 3.25 25 0.008380173
#> 111 2.75 2.75 25 0.014265745
#> 112 3.25 2.75 25 0.014508399
#> 113 2.75 3.25 25 0.014576489
#> 114 3.25 3.25 25 0.016045865
#> 115 3.75 2.75 25 0.031166747
#> 116 4.25 2.75 25 0.032291130
#> 117 3.75 3.25 25 0.036107783
#> 118 4.25 3.25 25 0.042671938
#> 119 4.75 2.75 25 0.047541474
#> 120 5.25 2.75 25 0.048043217
#> 121 4.75 3.25 25 0.062095064
#> 122 5.25 3.25 25 0.062376296
#> 123 5.75 2.75 25 0.033887058
#> 124 6.25 2.75 25 0.033616496
#> 125 5.75 3.25 25 0.044889800
#> 126 6.25 3.25 25 0.041046837
#> 127 6.75 2.75 25 0.017149671
#> 128 7.25 2.75 25 0.016663606
#> 129 6.75 3.25 25 0.021518455
#> 130 7.25 3.25 25 0.018393108
#> 131 7.75 2.75 25 0.010230146
#> 132 7.75 3.25 25 0.010262130
#> 133 8.25 2.75 25 0.007927758
#> 134 8.75 2.75 25 0.007261920
#> 135 8.25 3.25 25 0.007471795
#> 136 8.75 3.25 25 0.006967671
#> 137 9.25 2.75 25 0.006595192
#> 138 9.75 2.75 25 0.006484232
#> 139 9.25 3.25 25 0.006499175
#> 140 9.75 3.25 25 0.006465616
#> 141 0.25 3.75 25 0.006505738
#> 142 0.75 3.75 25 0.006505877
#> 143 0.25 4.25 25 0.006505757
#> 144 0.75 4.25 25 0.006506031
#> 145 1.25 3.75 25 0.007141714
#> 146 1.75 3.75 25 0.007157498
#> 147 1.25 4.25 25 0.007143420
#> 148 1.75 4.25 25 0.007182259
#> 149 2.25 3.75 25 0.010867797
#> 150 2.25 4.25 25 0.011153773
#> 151 2.75 3.75 25 0.031636941
#> 152 3.25 3.75 25 0.037396058
#> 153 2.75 4.25 25 0.033844018
#> 154 3.25 4.25 25 0.048454104
#> 155 3.75 3.75 25 0.154282455
#> 156 4.25 3.75 25 0.179818522
#> 157 3.75 4.25 25 0.190139988
#> 158 4.25 4.25 25 0.255255291
#> 159 4.75 3.75 25 0.345493950
#> 160 5.25 3.75 25 0.342918940
#> 161 4.75 4.25 25 0.449381935
#> 162 5.25 4.25 25 0.439487187
#> 163 5.75 3.75 25 0.179381783
#> 164 6.25 3.75 25 0.164065478
#> 165 5.75 4.25 25 0.246840064
#> 166 6.25 4.25 25 0.210202244
#> 167 6.75 3.75 25 0.053069265
#> 168 7.25 3.75 25 0.044497532
#> 169 6.75 4.25 25 0.089768713
#> 170 7.25 4.25 25 0.071359983
#> 171 7.75 3.75 25 0.017548941
#> 172 7.75 4.25 25 0.031814182
#> 173 8.25 3.75 25 0.009448275
#> 174 8.75 3.75 25 0.007660131
#> 175 8.25 4.25 25 0.014486377
#> 176 8.75 4.25 25 0.008813217
#> 177 9.25 3.75 25 0.006582913
#> 178 9.75 3.75 25 0.006513256
#> 179 9.25 4.25 25 0.006752308
#> 180 9.75 4.25 25 0.006529129
#> 181 0.25 4.75 25 0.006529402
#> 182 0.75 4.75 25 0.006529796
#> 183 0.25 5.25 25 0.006529405
#> 184 0.75 5.25 25 0.006529819
#> 185 1.25 4.75 25 0.007338933
#> 186 1.75 4.75 25 0.007401327
#> 187 1.25 5.25 25 0.007339150
#> 188 1.75 5.25 25 0.007404687
#> 189 2.25 4.75 25 0.012895115
#> 190 2.25 5.25 25 0.012939316
#> 191 2.75 4.75 25 0.050361815
#> 192 3.25 4.75 25 0.074426553
#> 193 2.75 5.25 25 0.050755837
#> 194 3.25 5.25 25 0.076709361
#> 195 3.75 4.75 25 0.375467203
#> 196 4.25 4.75 25 0.484281878
#> 197 3.75 5.25 25 0.383987620
#> 198 4.25 5.25 25 0.504706009
#> 199 4.75 4.75 25 0.995427856
#> 200 5.25 4.75 25 0.977888981
#> 201 4.75 5.25 25 1.026917312
#> 202 5.25 5.25 25 1.010196328
#> 203 5.75 4.75 25 0.469555993
#> 204 6.25 4.75 25 0.414290972
#> 205 5.75 5.25 25 0.496367556
#> 206 6.25 5.25 25 0.441584941
#> 207 6.75 4.75 25 0.156401887
#> 208 7.25 4.75 25 0.127116868
#> 209 6.75 5.25 25 0.187973141
#> 210 7.25 5.25 25 0.154722788
#> 211 7.75 4.75 25 0.056114335
#> 212 7.75 5.25 25 0.071873996
#> 213 8.25 4.75 25 0.022884097
#> 214 8.75 4.75 25 0.010903312
#> 215 8.25 5.25 25 0.028628172
#> 216 8.75 5.25 25 0.012239923
#> 217 9.25 4.75 25 0.007057267
#> 218 9.75 4.75 25 0.006579521
#> 219 9.25 5.25 25 0.007256205
#> 220 9.75 5.25 25 0.006598483
#> 221 0.25 5.75 25 0.006505763
#> 222 0.75 5.75 25 0.006506071
#> 223 0.25 6.25 25 0.006505745
#> 224 0.75 6.25 25 0.006505906
#> 225 1.25 5.75 25 0.007143854
#> 226 1.75 5.75 25 0.007189377
#> 227 1.25 6.25 25 0.007142074
#> 228 1.75 6.25 25 0.007164201
#> 229 2.25 5.75 25 0.011249354
#> 230 2.25 6.25 25 0.010964175
#> 231 2.75 5.75 25 0.034712616
#> 232 3.25 5.75 25 0.053603042
#> 233 2.75 6.25 25 0.032572815
#> 234 3.25 6.25 25 0.043364527
#> 235 3.75 5.75 25 0.209895261
#> 236 4.25 5.75 25 0.304209050
#> 237 3.75 6.25 25 0.179112049
#> 238 4.25 6.25 25 0.247052700
#> 239 4.75 5.75 25 0.527873803
#> 240 5.25 5.75 25 0.523064093
#> 241 4.75 6.25 25 0.463977482
#> 242 5.25 6.25 25 0.480085222
#> 243 5.75 5.75 25 0.315110242
#> 244 6.25 5.75 25 0.271700307
#> 245 5.75 6.25 25 0.289734547
#> 246 6.25 6.25 25 0.239424035
#> 247 6.75 5.75 25 0.154640400
#> 248 7.25 5.75 25 0.126579955
#> 249 6.75 6.25 25 0.109987597
#> 250 7.25 6.25 25 0.086616231
#> 251 7.75 5.75 25 0.063169500
#> 252 7.75 6.25 25 0.040623937
#> 253 8.25 5.75 25 0.025907287
#> 254 8.75 5.75 25 0.011471057
#> 255 8.25 6.25 25 0.017786306
#> 256 8.75 6.25 25 0.009594772
#> 257 9.25 5.75 25 0.007147901
#> 258 9.75 5.75 25 0.006566824
#> 259 9.25 6.25 25 0.006869675
#> 260 9.75 6.25 25 0.006540353
#> 261 0.25 6.75 25 0.006459835
#> 262 0.75 6.75 25 0.006459917
#> 263 0.25 7.25 25 0.006460537
#> 264 0.75 7.25 25 0.006460826
#> 265 1.25 6.75 25 0.006776713
#> 266 1.75 6.75 25 0.006786591
#> 267 1.25 7.25 25 0.006777183
#> 268 1.75 7.25 25 0.006783577
#> 269 2.25 6.75 25 0.008477554
#> 270 2.25 7.25 25 0.008450608
#> 271 2.75 6.75 25 0.015629707
#> 272 3.25 6.75 25 0.023474468
#> 273 2.75 7.25 25 0.015542837
#> 274 3.25 7.25 25 0.023947601
#> 275 3.75 6.75 25 0.070086208
#> 276 4.25 6.75 25 0.143199521
#> 277 3.75 7.25 25 0.076046884
#> 278 4.25 7.25 25 0.169439995
#> 279 4.75 6.75 25 0.254166234
#> 280 5.25 6.75 25 0.299503478
#> 281 4.75 7.25 25 0.316758010
#> 282 5.25 7.25 25 0.387406107
#> 283 5.75 6.75 25 0.236030209
#> 284 6.25 6.75 25 0.146871740
#> 285 5.75 7.25 25 0.308869401
#> 286 6.25 7.25 25 0.177753210
#> 287 6.75 6.75 25 0.068984344
#> 288 7.25 6.75 25 0.040201025
#> 289 6.75 7.25 25 0.067246297
#> 290 7.25 7.25 25 0.028903315
#> 291 7.75 6.75 25 0.019958480
#> 292 7.75 7.25 25 0.012228941
#> 293 8.25 6.75 25 0.010709058
#> 294 8.75 6.75 25 0.007671473
#> 295 8.25 7.25 25 0.007797277
#> 296 8.75 7.25 25 0.006990563
#> 297 9.25 6.75 25 0.006592162
#> 298 9.75 6.75 25 0.006472354
#> 299 9.25 7.25 25 0.006490831
#> 300 9.75 7.25 25 0.006462701
#> 301 0.25 7.75 25 0.006433408
#> 302 0.75 7.75 25 0.006435507
#> 303 1.25 7.75 25 0.006536546
#> 304 1.75 7.75 25 0.006539029
#> 305 2.25 7.75 25 0.007094433
#> 306 2.75 7.75 25 0.009540783
#> 307 3.25 7.75 25 0.017356967
#> 308 3.75 7.75 25 0.053932746
#> 309 4.25 7.75 25 0.143756479
#> 310 4.75 7.75 25 0.274915154
#> 311 5.25 7.75 25 0.345185556
#> 312 5.75 7.75 25 0.281684140
#> 313 6.25 7.75 25 0.152473860
#> 314 6.75 7.75 25 0.056358094
#> 315 7.25 7.75 25 0.019141460
#> 316 7.75 7.75 25 0.008758213
#> 317 8.25 7.75 25 0.006786852
#> 318 8.75 7.75 25 0.006567358
#> 319 9.25 7.75 25 0.006432291
#> 320 9.75 7.75 25 0.006427876
#> 321 0.25 8.25 25 0.006444591
#> 322 0.75 8.25 25 0.006454966
#> 323 0.25 8.75 25 0.006509169
#> 324 0.75 8.75 25 0.006542142
#> 325 1.25 8.25 25 0.006471051
#> 326 1.75 8.25 25 0.006459096
#> 327 1.25 8.75 25 0.006546436
#> 328 1.75 8.75 25 0.006498855
#> 329 2.25 8.25 25 0.006601210
#> 330 2.25 8.75 25 0.006577577
#> 331 2.75 8.25 25 0.007536681
#> 332 3.25 8.25 25 0.012464280
#> 333 2.75 8.75 25 0.007090432
#> 334 3.25 8.75 25 0.009097834
#> 335 3.75 8.25 25 0.034483368
#> 336 4.25 8.25 25 0.091462591
#> 337 3.75 8.75 25 0.018307116
#> 338 4.25 8.75 25 0.041600817
#> 339 4.75 8.25 25 0.174151710
#> 340 5.25 8.25 25 0.219044875
#> 341 4.75 8.75 25 0.075550388
#> 342 5.25 8.75 25 0.093955463
#> 343 5.75 8.25 25 0.179512366
#> 344 6.25 8.25 25 0.097197694
#> 345 5.75 8.75 25 0.077696799
#> 346 6.25 8.75 25 0.044007623
#> 347 6.75 8.25 25 0.036999302
#> 348 7.25 8.25 25 0.013318396
#> 349 6.75 8.75 25 0.019135578
#> 350 7.25 8.75 25 0.009436079
#> 351 7.75 8.25 25 0.007458037
#> 352 7.75 8.75 25 0.006897845
#> 353 8.25 8.25 25 0.006528102
#> 354 8.75 8.25 25 0.006443745
#> 355 8.25 8.75 25 0.006469488
#> 356 8.75 8.75 25 0.006438316
#> 357 9.25 8.25 25 0.006415563
#> 358 9.75 8.25 25 0.006415004
#> 359 9.25 8.75 25 0.006415052
#> 360 9.75 8.75 25 0.006414963
#> 361 0.25 9.25 25 0.006604126
#> 362 0.75 9.25 25 0.006671337
#> 363 0.25 9.75 25 0.006663095
#> 364 0.75 9.75 25 0.006750941
#> 365 1.25 9.25 25 0.006639383
#> 366 1.75 9.25 25 0.006539894
#> 367 1.25 9.75 25 0.006708274
#> 368 1.75 9.75 25 0.006577814
#> 369 2.25 9.25 25 0.006478707
#> 370 2.25 9.75 25 0.006486542
#> 371 2.75 9.25 25 0.006551009
#> 372 3.25 9.25 25 0.007069785
#> 373 2.75 9.75 25 0.006486646
#> 374 3.25 9.75 25 0.006564082
#> 375 3.75 9.25 25 0.009434594
#> 376 4.25 9.25 25 0.015547543
#> 377 3.75 9.75 25 0.006998706
#> 378 4.25 9.75 25 0.008027661
#> 379 4.75 9.25 25 0.024435410
#> 380 5.25 9.25 25 0.029275710
#> 381 4.75 9.75 25 0.009544433
#> 382 5.25 9.75 25 0.010360900
#> 383 5.75 9.25 25 0.025040717
#> 384 6.25 9.25 25 0.016193969
#> 385 5.75 9.75 25 0.009629059
#> 386 6.25 9.75 25 0.008138896
#> 387 6.75 9.25 25 0.009708253
#> 388 7.25 9.25 25 0.007158546
#> 389 6.75 9.75 25 0.007007561
#> 390 7.25 9.75 25 0.006577628
#> 391 7.75 9.25 25 0.006523903
#> 392 7.75 9.75 25 0.006443701
#> 393 8.25 9.25 25 0.006423519
#> 394 8.75 9.25 25 0.006415450
#> 395 8.25 9.75 25 0.006416399
#> 396 8.75 9.75 25 0.006415040
#> 397 9.25 9.25 25 0.006412095
#> 398 9.75 9.25 25 0.006412077
#> 399 9.25 9.75 25 0.006412079
#> 400 9.75 9.75 25 0.006412076