Compute the conditional intensity at a location

Compute the conditional intensity at a location

Usage

temporal_conditional_intensity(hawkes, parameters, coordinates, step = 0.1)

Arguments

hawkes

A hawkes object.

parameters

Named list containing background, triggering, spatial, and temporal parameters.

coordinates

Numeric vector of length two giving the evaluation location.

step

Grid step size used to build the temporal evaluation grid.

Value

A numeric vector.

Examples

spatial_region <- create_rectangular_sf(0, 10, 0, 10)

params <- list(
  background_rate = list(intercept = -4),
  triggering_rate = 0.75,
  spatial = list(mean = 0, sd = 0.75),
  temporal = list(rate = 2)
)
hawkes <- rHawkes(params, time_window = c(0, 50), spatial_region = spatial_region)
temporal_conditional_intensity(hawkes, params, c(5, 5))
#>     x y            t  intensity
#> 1   5 5  0.000000000 0.01831564
#> 2   5 5  0.007962508 0.01895313
#> 3   5 5  0.100000000 0.01884595
#> 4   5 5  0.200000000 0.01874982
#> 5   5 5  0.300000000 0.01867112
#> 6   5 5  0.400000000 0.01860668
#> 7   5 5  0.500000000 0.01855392
#> 8   5 5  0.600000000 0.01851073
#> 9   5 5  0.700000000 0.01847537
#> 10  5 5  0.800000000 0.01844641
#> 11  5 5  0.886506693 0.01842564
#> 12  5 5  0.897211119 0.01860663
#> 13  5 5  0.900000000 0.01860501
#> 14  5 5  1.000000000 0.01855256
#> 15  5 5  1.100000000 0.01850961
#> 16  5 5  1.103803635 0.01850841
#> 17  5 5  1.200000000 0.01847467
#> 18  5 5  1.300000000 0.01844584
#> 19  5 5  1.400000000 0.01842224
#> 20  5 5  1.500000000 0.01840292
#> 21  5 5  1.600000000 0.01838710
#> 22  5 5  1.700000000 0.01837414
#> 23  5 5  1.800000000 0.01836354
#> 24  5 5  1.846987486 0.01835925
#> 25  5 5  1.900000000 0.01835486
#> 26  5 5  2.000000000 0.01834775
#> 27  5 5  2.100000000 0.01834193
#> 28  5 5  2.200000000 0.01833717
#> 29  5 5  2.300000000 0.01833326
#> 30  5 5  2.400000000 0.01833007
#> 31  5 5  2.500000000 0.01832745
#> 32  5 5  2.600000000 0.01832531
#> 33  5 5  2.700000000 0.01832356
#> 34  5 5  2.800000000 0.01832212
#> 35  5 5  2.900000000 0.01832095
#> 36  5 5  2.923920099 0.01832070
#> 37  5 5  3.000000000 0.01831999
#> 38  5 5  3.090292916 0.01831927
#> 39  5 5  3.100000000 0.01831920
#> 40  5 5  3.200000000 0.01831855
#> 41  5 5  3.300000000 0.01831802
#> 42  5 5  3.400000000 0.01831759
#> 43  5 5  3.500000000 0.01831724
#> 44  5 5  3.600000000 0.01831695
#> 45  5 5  3.700000000 0.01831671
#> 46  5 5  3.800000000 0.01831652
#> 47  5 5  3.808539134 0.01831651
#> 48  5 5  3.900000000 0.01831636
#> 49  5 5  4.000000000 0.01831623
#> 50  5 5  4.100000000 0.01831613
#> 51  5 5  4.200000000 0.01831604
#> 52  5 5  4.300000000 0.01831596
#> 53  5 5  4.400000000 0.01831591
#> 54  5 5  4.500000000 0.01831586
#> 55  5 5  4.600000000 0.01831582
#> 56  5 5  4.700000000 0.01831579
#> 57  5 5  4.731387441 0.01831578
#> 58  5 5  4.800000000 0.01831576
#> 59  5 5  4.900000000 0.01831574
#> 60  5 5  5.000000000 0.01831572
#> 61  5 5  5.100000000 0.01831570
#> 62  5 5  5.100927680 0.01831571
#> 63  5 5  5.180145219 0.01831580
#> 64  5 5  5.200000000 0.01831580
#> 65  5 5  5.300000000 0.01831577
#> 66  5 5  5.400000000 0.01831574
#> 67  5 5  5.500000000 0.01831573
#> 68  5 5  5.545841688 0.01831585
#> 69  5 5  5.600000000 0.01831583
#> 70  5 5  5.700000000 0.01831579
#> 71  5 5  5.800000000 0.01831577
#> 72  5 5  5.900000000 0.01831574
#> 73  5 5  5.966085671 0.01831574
#> 74  5 5  6.000000000 0.01831574
#> 75  5 5  6.066782657 0.01889259
#> 76  5 5  6.100000000 0.01885550
#> 77  5 5  6.111278139 0.01919986
#> 78  5 5  6.144017382 0.01921563
#> 79  5 5  6.147111899 0.01921030
#> 80  5 5  6.200000000 0.01912050
#> 81  5 5  6.218263040 0.01992362
#> 82  5 5  6.269027184 0.01976923
#> 83  5 5  6.300000000 0.01968192
#> 84  5 5  6.400000000 0.01943425
#> 85  5 5  6.500000000 0.01923148
#> 86  5 5  6.523173457 0.01919001
#> 87  5 5  6.528071947 0.01918156
#> 88  5 5  6.600000000 0.01906553
#> 89  5 5  6.700000000 0.01892960
#> 90  5 5  6.782765236 0.02038321
#> 91  5 5  6.800000000 0.02031316
#> 92  5 5  6.855230355 0.02010427
#> 93  5 5  6.900000000 0.01995108
#> 94  5 5  6.903102517 0.01994096
#> 95  5 5  6.919436402 0.01988873
#> 96  5 5  6.954922685 0.01978096
#> 97  5 5  7.000000000 0.01965463
#> 98  5 5  7.042304790 0.02441318
#> 99  5 5  7.100000000 0.02374866
#> 100 5 5  7.200000000 0.02276382
#> 101 5 5  7.300000000 0.02195750
#> 102 5 5  7.400000000 0.02129734
#> 103 5 5  7.500000000 0.02075685
#> 104 5 5  7.600000000 0.02031434
#> 105 5 5  7.665946602 0.02006737
#> 106 5 5  7.700000000 0.01995203
#> 107 5 5  7.800000000 0.01965541
#> 108 5 5  7.802899255 0.01980291
#> 109 5 5  7.837777439 0.01970270
#> 110 5 5  7.900000000 0.01954039
#> 111 5 5  8.000000000 0.01931838
#> 112 5 5  8.100000000 0.01913662
#> 113 5 5  8.200000000 0.01898780
#> 114 5 5  8.287948706 0.01887939
#> 115 5 5  8.300000000 0.01886597
#> 116 5 5  8.400000000 0.01876621
#> 117 5 5  8.500000000 0.01868453
#> 118 5 5  8.600000000 0.01861766
#> 119 5 5  8.700000000 0.01856292
#> 120 5 5  8.800000000 0.01851809
#> 121 5 5  8.900000000 0.01848139
#> 122 5 5  9.000000000 0.01845135
#> 123 5 5  9.071477550 0.02278028
#> 124 5 5  9.100000000 0.02253272
#> 125 5 5  9.200000000 0.02176829
#> 126 5 5  9.300000000 0.02114243
#> 127 5 5  9.342120124 0.02111403
#> 128 5 5  9.400000000 0.02080814
#> 129 5 5  9.417102782 0.02072576
#> 130 5 5  9.439635973 0.02061995
#> 131 5 5  9.500000000 0.02035789
#> 132 5 5  9.600000000 0.01998769
#> 133 5 5  9.700000000 0.01968460
#> 134 5 5  9.800000000 0.01943645
#> 135 5 5  9.900000000 0.01923328
#> 136 5 5 10.000000000 0.01906694
#> 137 5 5 10.007341665 0.01905599
#> 138 5 5 10.100000000 0.01893075
#> 139 5 5 10.200000000 0.01881925
#> 140 5 5 10.210005371 0.01880927
#> 141 5 5 10.300000000 0.01872796
#> 142 5 5 10.400000000 0.01865322
#> 143 5 5 10.500000000 0.01859203
#> 144 5 5 10.600000000 0.01854193
#> 145 5 5 10.700000000 0.01850091
#> 146 5 5 10.708671934 0.02718582
#> 147 5 5 10.800000000 0.02570499
#> 148 5 5 10.900000000 0.02436553
#> 149 5 5 11.000000000 0.02326887
#> 150 5 5 11.100000000 0.02237100
#> 151 5 5 11.124830570 0.02217452
#> 152 5 5 11.200000000 0.02163589
#> 153 5 5 11.300000000 0.02103403
#> 154 5 5 11.400000000 0.02054127
#> 155 5 5 11.500000000 0.02013783
#> 156 5 5 11.564783030 0.01991640
#> 157 5 5 11.600000000 0.01980753
#> 158 5 5 11.700000000 0.01953710
#> 159 5 5 11.800000000 0.01931568
#> 160 5 5 11.900000000 0.01913441
#> 161 5 5 11.928025001 0.01908978
#> 162 5 5 12.000000000 0.01898599
#> 163 5 5 12.093346134 0.01887183
#> 164 5 5 12.100000000 0.01886447
#> 165 5 5 12.179560027 0.02232228
#> 166 5 5 12.200000000 0.02216179
#> 167 5 5 12.300000000 0.02146460
#> 168 5 5 12.384587099 0.02143166
#> 169 5 5 12.400000000 0.02133707
#> 170 5 5 12.500000000 0.02078938
#> 171 5 5 12.533919312 0.03951934
#> 172 5 5 12.600000000 0.03689431
#> 173 5 5 12.700000000 0.03352657
#> 174 5 5 12.800000000 0.03076929
#> 175 5 5 12.900000000 0.02851183
#> 176 5 5 13.000000000 0.02666357
#> 177 5 5 13.100000000 0.02515035
#> 178 5 5 13.200000000 0.02391143
#> 179 5 5 13.300000000 0.02289708
#> 180 5 5 13.400000000 0.02206661
#> 181 5 5 13.500000000 0.02138667
#> 182 5 5 13.600000000 0.02082999
#> 183 5 5 13.700000000 0.02037421
#> 184 5 5 13.713907376 0.02031774
#> 185 5 5 13.800000000 0.02000106
#> 186 5 5 13.803115557 0.05501107
#> 187 5 5 13.900000000 0.04854711
#> 188 5 5 14.000000000 0.04306707
#> 189 5 5 14.100000000 0.03858040
#> 190 5 5 14.111413830 0.03812304
#> 191 5 5 14.200000000 0.03490702
#> 192 5 5 14.300000000 0.03189951
#> 193 5 5 14.371066390 0.03009973
#> 194 5 5 14.400000000 0.02943718
#> 195 5 5 14.482384855 0.02864816
#> 196 5 5 14.500000000 0.02829048
#> 197 5 5 14.600000000 0.02648235
#> 198 5 5 14.700000000 0.02500198
#> 199 5 5 14.800000000 0.02378995
#> 200 5 5 14.900000000 0.02279763
#> 201 5 5 15.000000000 0.02198518
#> 202 5 5 15.100000000 0.02132000
#> 203 5 5 15.200000000 0.02077541
#> 204 5 5 15.300000000 0.02032953
#> 205 5 5 15.400000000 0.01996447
#> 206 5 5 15.487653442 0.01969937
#> 207 5 5 15.500000000 0.01966562
#> 208 5 5 15.583007801 0.02020734
#> 209 5 5 15.600000000 0.02014414
#> 210 5 5 15.647907822 0.01997707
#> 211 5 5 15.653489339 0.01996080
#> 212 5 5 15.700000000 0.01981466
#> 213 5 5 15.800000000 0.01954294
#> 214 5 5 15.868322447 0.01939407
#> 215 5 5 15.900000000 0.01932786
#> 216 5 5 16.000000000 0.01914438
#> 217 5 5 16.026995490 0.01910082
#> 218 5 5 16.100000000 0.01899415
#> 219 5 5 16.102769515 0.01899045
#> 220 5 5 16.169796951 0.01890579
#> 221 5 5 16.200000000 0.01887119
#> 222 5 5 16.266802793 0.02154114
#> 223 5 5 16.300000000 0.02133394
#> 224 5 5 16.400000000 0.02078682
#> 225 5 5 16.451463774 0.02054523
#> 226 5 5 16.500000000 0.02033898
#> 227 5 5 16.577359596 0.02004894
#> 228 5 5 16.600000000 0.01997221
#> 229 5 5 16.700000000 0.01967192
#> 230 5 5 16.720911388 0.01961637
#> 231 5 5 16.747338047 0.01954941
#> 232 5 5 16.800000000 0.01942607
#> 233 5 5 16.801389453 0.01942299
#> 234 5 5 16.865102650 0.01929276
#> 235 5 5 16.900000000 0.01922689
#> 236 5 5 17.000000000 0.01906171
#> 237 5 5 17.007633925 0.01905040
#> 238 5 5 17.062665773 0.01897382
#> 239 5 5 17.100000000 0.01892647
#> 240 5 5 17.200000000 0.01881574
#> 241 5 5 17.210168263 0.01882736
#> 242 5 5 17.227556367 0.01880993
#> 243 5 5 17.274790129 0.01876557
#> 244 5 5 17.296778044 0.01874655
#> 245 5 5 17.300000000 0.01874378
#> 246 5 5 17.329009649 0.01902432
#> 247 5 5 17.353567367 0.01899035
#> 248 5 5 17.366618278 0.01902477
#> 249 5 5 17.400000000 0.01897897
#> 250 5 5 17.470658560 0.01889156
#> 251 5 5 17.492767725 0.01886665
#> 252 5 5 17.500000000 0.01885874
#> 253 5 5 17.504461066 0.01885394
#> 254 5 5 17.600000000 0.01876031
#> 255 5 5 17.684486328 0.01869119
#> 256 5 5 17.700000000 0.01867972
#> 257 5 5 17.711625706 0.01867139
#> 258 5 5 17.712920945 0.01868264
#> 259 5 5 17.784035006 0.01863399
#> 260 5 5 17.800000000 0.01862398
#> 261 5 5 17.872989949 0.01858211
#> 262 5 5 17.900000000 0.01856809
#> 263 5 5 17.913310004 0.01856146
#> 264 5 5 18.000000000 0.01852233
#> 265 5 5 18.008565752 0.01854284
#> 266 5 5 18.100000000 0.01850487
#> 267 5 5 18.119178495 0.01849897
#> 268 5 5 18.149071287 0.01850086
#> 269 5 5 18.200000000 0.01848293
#> 270 5 5 18.225053187 0.02053050
#> 271 5 5 18.300000000 0.02022219
#> 272 5 5 18.358452406 0.02126321
#> 273 5 5 18.362538397 0.02450647
#> 274 5 5 18.400000000 0.02405958
#> 275 5 5 18.500000000 0.02301838
#> 276 5 5 18.600000000 0.02216592
#> 277 5 5 18.700000000 0.02146798
#> 278 5 5 18.700812822 0.02267100
#> 279 5 5 18.719063813 0.02253041
#> 280 5 5 18.800000000 0.02190051
#> 281 5 5 18.867072064 0.02145048
#> 282 5 5 18.900000000 0.02125068
#> 283 5 5 19.000000000 0.02071865
#> 284 5 5 19.100000000 0.02028306
#> 285 5 5 19.200000000 0.01992643
#> 286 5 5 19.300000000 0.01963444
#> 287 5 5 19.400000000 0.01939538
#> 288 5 5 19.500000000 0.01919966
#> 289 5 5 19.532740813 0.01914365
#> 290 5 5 19.571595123 0.01908176
#> 291 5 5 19.600000000 0.01903945
#> 292 5 5 19.614867272 0.02041873
#> 293 5 5 19.639075096 0.02032007
#> 294 5 5 19.696416438 0.02010290
#> 295 5 5 19.700000000 0.02009014
#> 296 5 5 19.714260572 0.02005738
#> 297 5 5 19.730000091 0.02057082
#> 298 5 5 19.782453157 0.02034633
#> 299 5 5 19.800000000 0.02027630
#> 300 5 5 19.848779758 0.02010203
#> 301 5 5 19.900000000 0.01992809
#> 302 5 5 20.000000000 0.01963580
#> 303 5 5 20.100000000 0.01939650
#> 304 5 5 20.200000000 0.01920057
#> 305 5 5 20.300000000 0.01904016
#> 306 5 5 20.311302715 0.02101109
#> 307 5 5 20.336348923 0.27021786
#> 308 5 5 20.400000000 0.24010738
#> 309 5 5 20.457937667 0.21616897
#> 310 5 5 20.500000000 0.20020550
#> 311 5 5 20.566757670 0.49091572
#> 312 5 5 20.600000000 0.46051680
#> 313 5 5 20.700000000 0.38035932
#> 314 5 5 20.800000000 0.31473194
#> 315 5 5 20.900000000 0.26100078
#> 316 5 5 20.918459169 0.31739582
#> 317 5 5 21.000000000 0.27239078
#> 318 5 5 21.100000000 0.22633477
#> 319 5 5 21.200000000 0.18862730
#> 320 5 5 21.221726315 0.19375425
#> 321 5 5 21.300000000 0.16833162
#> 322 5 5 21.388356909 0.34550613
#> 323 5 5 21.400000000 0.33797514
#> 324 5 5 21.414261102 0.32898661
#> 325 5 5 21.432232574 0.31801899
#> 326 5 5 21.462858536 0.66453842
#> 327 5 5 21.500000000 0.61827467
#> 328 5 5 21.563802387 0.54640029
#> 329 5 5 21.600000000 0.50952055
#> 330 5 5 21.700000000 0.42048020
#> 331 5 5 21.800000000 0.34758014
#> 332 5 5 21.900000000 0.28789461
#> 333 5 5 22.000000000 0.23902823
#> 334 5 5 22.100000000 0.19901983
#> 335 5 5 22.200000000 0.16626371
#> 336 5 5 22.300000000 0.13944528
#> 337 5 5 22.400000000 0.11748820
#> 338 5 5 22.465293852 0.12583584
#> 339 5 5 22.500000000 0.11862575
#> 340 5 5 22.544737106 0.11221103
#> 341 5 5 22.600000000 0.10238612
#> 342 5 5 22.700000000 0.08714673
#> 343 5 5 22.772340135 0.07801387
#> 344 5 5 22.800000000 0.07480107
#> 345 5 5 22.900000000 0.06456200
#> 346 5 5 22.937867068 0.06174528
#> 347 5 5 23.000000000 0.05667031
#> 348 5 5 23.026693295 0.05568000
#> 349 5 5 23.100000000 0.05058454
#> 350 5 5 23.198813385 0.04536773
#> 351 5 5 23.200000000 0.04530361
#> 352 5 5 23.300000000 0.04041152
#> 353 5 5 23.400000000 0.03640622
#> 354 5 5 23.497635834 0.03320558
#> 355 5 5 23.500000000 0.03313534
#> 356 5 5 23.505872029 0.03301106
#> 357 5 5 23.600000000 0.03048936
#> 358 5 5 23.700000000 0.02828264
#> 359 5 5 23.754820083 0.02728198
#> 360 5 5 23.800000000 0.02650731
#> 361 5 5 23.900000000 0.02502241
#> 362 5 5 23.906525134 0.02493545
#> 363 5 5 23.966538524 0.02419835
#> 364 5 5 24.000000000 0.02381754
#> 365 5 5 24.015004085 0.02365489
#> 366 5 5 24.057182500 0.02322297
#> 367 5 5 24.100000000 0.02282022
#> 368 5 5 24.162665164 0.02229392
#> 369 5 5 24.194578505 0.02212571
#> 370 5 5 24.200000000 0.02208462
#> 371 5 5 24.215540238 0.04716044
#> 372 5 5 24.234924016 0.05623237
#> 373 5 5 24.264431291 0.05407078
#> 374 5 5 24.272160785 0.05352238
#> 375 5 5 24.300000000 0.05161570
#> 376 5 5 24.348657019 0.04852782
#> 377 5 5 24.400000000 0.04557942
#> 378 5 5 24.419883740 0.04480137
#> 379 5 5 24.485621892 0.04153843
#> 380 5 5 24.500000000 0.04088014
#> 381 5 5 24.571243550 0.03788355
#> 382 5 5 24.600000000 0.03678989
#> 383 5 5 24.700000000 0.03344108
#> 384 5 5 24.736227835 0.03238392
#> 385 5 5 24.785089838 0.03107669
#> 386 5 5 24.800000000 0.03070177
#> 387 5 5 24.900000000 0.02845655
#> 388 5 5 24.973536584 0.02830190
#> 389 5 5 25.000000000 0.02778710
#> 390 5 5 25.100000000 0.02607022
#> 391 5 5 25.200000000 0.02466455
#> 392 5 5 25.300000000 0.02351369
#> 393 5 5 25.400000000 0.02257144
#> 394 5 5 25.500000000 0.02180000
#> 395 5 5 25.600000000 0.02116839
#> 396 5 5 25.700000000 0.02065127
#> 397 5 5 25.710492839 0.02539772
#> 398 5 5 25.800000000 0.02423693
#> 399 5 5 25.900000000 0.02316358
#> 400 5 5 26.000000000 0.02228480
#> 401 5 5 26.012622082 0.02218594
#> 402 5 5 26.100000000 0.02156539
#> 403 5 5 26.200000000 0.02097631
#> 404 5 5 26.230658258 0.02081807
#> 405 5 5 26.300000000 0.02049401
#> 406 5 5 26.400000000 0.02009914
#> 407 5 5 26.500000000 0.01977585
#> 408 5 5 26.600000000 0.01951115
#> 409 5 5 26.700000000 0.01929444
#> 410 5 5 26.800000000 0.01911702
#> 411 5 5 26.900000000 0.01897175
#> 412 5 5 27.000000000 0.01885282
#> 413 5 5 27.033057982 0.01881849
#> 414 5 5 27.100000000 0.01875548
#> 415 5 5 27.136734945 0.02584948
#> 416 5 5 27.176582281 0.22862463
#> 417 5 5 27.200000000 0.21900182
#> 418 5 5 27.218850225 0.21157671
#> 419 5 5 27.300000000 0.18262359
#> 420 5 5 27.400000000 0.15283961
#> 421 5 5 27.500000000 0.12845455
#> 422 5 5 27.600000000 0.10848975
#> 423 5 5 27.700000000 0.09214396
#> 424 5 5 27.800000000 0.07876116
#> 425 5 5 27.900000000 0.06780424
#> 426 5 5 27.982773003 0.06025381
#> 427 5 5 28.000000000 0.05883348
#> 428 5 5 28.100000000 0.05148884
#> 429 5 5 28.123129026 0.05048111
#> 430 5 5 28.200000000 0.04589732
#> 431 5 5 28.232600139 0.04415636
#> 432 5 5 28.300000000 0.04089761
#> 433 5 5 28.400000000 0.03680419
#> 434 5 5 28.500000000 0.03345279
#> 435 5 5 28.600000000 0.03070889
#> 436 5 5 28.700000000 0.02846237
#> 437 5 5 28.767573898 0.02717969
#> 438 5 5 28.800000000 0.02662308
#> 439 5 5 28.813284368 0.27086718
#> 440 5 5 28.900000000 0.23065465
#> 441 5 5 28.927478757 0.21929990
#> 442 5 5 28.940511830 0.21412871
#> 443 5 5 28.944721226 0.21248712
#> 444 5 5 29.000000000 0.19216413
#> 445 5 5 29.077745146 0.16713843
#> 446 5 5 29.100000000 0.16065963
#> 447 5 5 29.200000000 0.13485704
#> 448 5 5 29.212242275 0.13203824
#> 449 5 5 29.300000000 0.11373168
#> 450 5 5 29.311595038 0.11154454
#> 451 5 5 29.400000000 0.09643577
#> 452 5 5 29.424566739 0.09271662
#> 453 5 5 29.500000000 0.08229769
#> 454 5 5 29.545845187 0.29159758
#> 455 5 5 29.546992200 0.29243162
#> 456 5 5 29.600000000 0.26485846
#> 457 5 5 29.697334488 0.22274800
#> 458 5 5 29.700000000 0.22166106
#> 459 5 5 29.800000000 0.18480079
#> 460 5 5 29.854528226 0.16759951
#> 461 5 5 29.886831669 0.15825970
#> 462 5 5 29.900000000 0.15462215
#> 463 5 5 30.000000000 0.12991397
#> 464 5 5 30.100000000 0.10968463
#> 465 5 5 30.200000000 0.09312224
#> 466 5 5 30.300000000 0.07956210
#> 467 5 5 30.320572146 0.07709335
#> 468 5 5 30.371281312 0.07142460
#> 469 5 5 30.377134170 0.07080774
#> 470 5 5 30.385380332 0.06995248
#> 471 5 5 30.400000000 0.06846452
#> 472 5 5 30.500000000 0.05937407
#> 473 5 5 30.531276788 0.05690922
#> 474 5 5 30.600000000 0.05195308
#> 475 5 5 30.700000000 0.04585564
#> 476 5 5 30.800000000 0.04086349
#> 477 5 5 30.861881611 0.03823867
#> 478 5 5 30.900000000 0.03677626
#> 479 5 5 31.000000000 0.03342991
#> 480 5 5 31.083856830 0.03109663
#> 481 5 5 31.100000000 0.03069057
#> 482 5 5 31.170170963 0.02907306
#> 483 5 5 31.200000000 0.02845006
#> 484 5 5 31.211284776 0.02822389
#> 485 5 5 31.219108908 0.02807006
#> 486 5 5 31.300000000 0.02661300
#> 487 5 5 31.400000000 0.02510895
#> 488 5 5 31.438555373 0.02460480
#> 489 5 5 31.457042689 0.02437651
#> 490 5 5 31.500000000 0.02387753
#> 491 5 5 31.507120490 0.02380001
#> 492 5 5 31.516710609 0.02634278
#> 493 5 5 31.559648572 0.02829115
#> 494 5 5 31.600000000 0.02751772
#> 495 5 5 31.698812099 0.02586759
#> 496 5 5 31.700000000 0.02584967
#> 497 5 5 31.752189090 0.02510312
#> 498 5 5 31.800000000 0.02448415
#> 499 5 5 31.851919153 0.02387576
#> 500 5 5 31.900000000 0.02336599
#> 501 5 5 32.000000000 0.02245052
#> 502 5 5 32.100000000 0.02170099
#> 503 5 5 32.200000000 0.02108733
#> 504 5 5 32.204015871 0.02113249
#> 505 5 5 32.208721835 0.02111549
#> 506 5 5 32.243379867 0.02093167
#> 507 5 5 32.300000000 0.02065159
#> 508 5 5 32.315873962 0.02905809
#> 509 5 5 32.400000000 0.02739452
#> 510 5 5 32.500000000 0.02574880
#> 511 5 5 32.509457463 0.02560952
#> 512 5 5 32.600000000 0.02440139
#> 513 5 5 32.700000000 0.02329823
#> 514 5 5 32.733637477 0.02313756
#> 515 5 5 32.766317927 0.02284987
#> 516 5 5 32.800000000 0.02255449
#> 517 5 5 32.800085593 0.02255694
#> 518 5 5 32.900000000 0.02178872
#> 519 5 5 33.000000000 0.02115916
#> 520 5 5 33.028830688 0.02168571
#> 521 5 5 33.063510305 0.02146011
#> 522 5 5 33.100000000 0.02123880
#> 523 5 5 33.156983463 0.02092394
#> 524 5 5 33.180990192 0.02080167
#> 525 5 5 33.200000000 0.02070892
#> 526 5 5 33.300000000 0.02027509
#> 527 5 5 33.340015196 0.02012439
#> 528 5 5 33.400000000 0.01991991
#> 529 5 5 33.500000000 0.01962910
#> 530 5 5 33.600000000 0.01939101
#> 531 5 5 33.669127463 0.01927219
#> 532 5 5 33.700000000 0.01921491
#> 533 5 5 33.800000000 0.01905190
#> 534 5 5 33.900000000 0.01891844
#> 535 5 5 34.000000000 0.01880917
#> 536 5 5 34.100000000 0.01871971
#> 537 5 5 34.162124456 0.01869089
#> 538 5 5 34.200000000 0.01866351
#> 539 5 5 34.251985304 0.01865336
#> 540 5 5 34.300000000 0.01862244
#> 541 5 5 34.400000000 0.01856683
#> 542 5 5 34.500000000 0.01852129
#> 543 5 5 34.600000000 0.01848402
#> 544 5 5 34.610193094 0.01848062
#> 545 5 5 34.700000000 0.01845349
#> 546 5 5 34.800000000 0.01842851
#> 547 5 5 34.813328866 0.01849844
#> 548 5 5 34.900000000 0.01846935
#> 549 5 5 35.000000000 0.01844149
#> 550 5 5 35.100000000 0.01841867
#> 551 5 5 35.119612608 0.07831498
#> 552 5 5 35.200000000 0.06940411
#> 553 5 5 35.300000000 0.06014334
#> 554 5 5 35.400000000 0.05256126
#> 555 5 5 35.500000000 0.04635359
#> 556 5 5 35.600000000 0.04127117
#> 557 5 5 35.610413793 0.04079936
#> 558 5 5 35.700000000 0.03711117
#> 559 5 5 35.800000000 0.03370412
#> 560 5 5 35.900000000 0.03091466
#> 561 5 5 36.000000000 0.02863085
#> 562 5 5 36.100000000 0.02676102
#> 563 5 5 36.200000000 0.02523013
#> 564 5 5 36.300000000 0.02397674
#> 565 5 5 36.400000000 0.02295056
#> 566 5 5 36.500000000 0.02211039
#> 567 5 5 36.600000000 0.02142252
#> 568 5 5 36.700000000 0.02085934
#> 569 5 5 36.800000000 0.02039824
#> 570 5 5 36.900000000 0.02002073
#> 571 5 5 36.959273235 0.02048616
#> 572 5 5 37.000000000 0.02031637
#> 573 5 5 37.020859308 0.02023462
#> 574 5 5 37.100000000 0.01995370
#> 575 5 5 37.190642607 0.01968210
#> 576 5 5 37.200000000 0.01965677
#> 577 5 5 37.235864336 0.01956420
#> 578 5 5 37.251786248 0.01952507
#> 579 5 5 37.264031674 0.01949819
#> 580 5 5 37.300000000 0.01941611
#> 581 5 5 37.361375490 0.01928899
#> 582 5 5 37.377105795 0.02316147
#> 583 5 5 37.400000000 0.02294459
#> 584 5 5 37.500000000 0.02210550
#> 585 5 5 37.600000000 0.02141852
#> 586 5 5 37.700000000 0.02085606
#> 587 5 5 37.708897096 0.02081126
#> 588 5 5 37.800000000 0.02039556
#> 589 5 5 37.900000000 0.02001854
#> 590 5 5 37.941743161 0.01988223
#> 591 5 5 38.000000000 0.01970994
#> 592 5 5 38.100000000 0.01945719
#> 593 5 5 38.200000000 0.01925026
#> 594 5 5 38.227508441 0.01920023
#> 595 5 5 38.300000000 0.01908085
#> 596 5 5 38.362933199 0.01899035
#> 597 5 5 38.400000000 0.01894214
#> 598 5 5 38.407661025 0.01893261
#> 599 5 5 38.456707883 0.01887538
#> 600 5 5 38.500000000 0.01882895
#> 601 5 5 38.516582618 0.01881221
#> 602 5 5 38.564683244 0.01876667
#> 603 5 5 38.572336873 0.01875985
#> 604 5 5 38.600000000 0.01873594
#> 605 5 5 38.656979124 0.01869067
#> 606 5 5 38.683250275 0.01867147
#> 607 5 5 38.700000000 0.01865975
#> 608 5 5 38.785712320 0.01860554
#> 609 5 5 38.800000000 0.01859737
#> 610 5 5 38.870656012 0.01856033
#> 611 5 5 38.900000000 0.01854639
#> 612 5 5 38.908942846 0.01865799
#> 613 5 5 38.987368053 0.01860829
#> 614 5 5 39.000000000 0.01860099
#> 615 5 5 39.018677390 0.02379220
#> 616 5 5 39.100000000 0.02297013
#> 617 5 5 39.200000000 0.02212641
#> 618 5 5 39.212958307 0.02202892
#> 619 5 5 39.266608346 0.02165113
#> 620 5 5 39.294615464 0.02146943
#> 621 5 5 39.300000000 0.02143565
#> 622 5 5 39.400000000 0.02087009
#> 623 5 5 39.432229011 0.02071063
#> 624 5 5 39.484508126 0.02047292
#> 625 5 5 39.500000000 0.02040710
#> 626 5 5 39.593958668 0.02004880
#> 627 5 5 39.600000000 0.02002799
#> 628 5 5 39.652772840 0.01985852
#> 629 5 5 39.694809082 0.01973411
#> 630 5 5 39.700000000 0.01971946
#> 631 5 5 39.735349595 0.01962364
#> 632 5 5 39.742546380 0.01960495
#> 633 5 5 39.767118828 0.01954909
#> 634 5 5 39.800000000 0.01947058
#> 635 5 5 39.867178277 0.01932608
#> 636 5 5 39.900000000 0.01926188
#> 637 5 5 39.966819949 0.01914351
#> 638 5 5 40.000000000 0.01909035
#> 639 5 5 40.100000000 0.01894992
#> 640 5 5 40.118366110 0.01913303
#> 641 5 5 40.183236769 0.01903390
#> 642 5 5 40.200000000 0.01901022
#> 643 5 5 40.300000000 0.01888431
#> 644 5 5 40.319656635 0.01889438
#> 645 5 5 40.332849661 0.01888071
#> 646 5 5 40.360123635 0.01885071
#> 647 5 5 40.384735411 0.01882501
#> 648 5 5 40.400000000 0.01880970
#> 649 5 5 40.500000000 0.01872014
#> 650 5 5 40.547023546 0.01931226
#> 651 5 5 40.600000000 0.01921207
#> 652 5 5 40.631692588 0.01916758
#> 653 5 5 40.695292393 0.01906583
#> 654 5 5 40.700000000 0.01905880
#> 655 5 5 40.714412832 0.01903769
#> 656 5 5 40.720179958 0.01918147
#> 657 5 5 40.736734869 0.01915328
#> 658 5 5 40.746885595 0.01913645
#> 659 5 5 40.800000000 0.01905373
#> 660 5 5 40.900000000 0.01891994
#> 661 5 5 40.974313478 0.01883648
#> 662 5 5 41.000000000 0.01881039
#> 663 5 5 41.077835615 0.01873923
#> 664 5 5 41.081293202 0.01941621
#> 665 5 5 41.100000000 0.01937580
#> 666 5 5 41.132381364 0.01930944
#> 667 5 5 41.153520990 0.02005131
#> 668 5 5 41.165359684 0.02001069
#> 669 5 5 41.165590032 0.02000991
#> 670 5 5 41.200000000 0.01989723
#> 671 5 5 41.201092717 0.01989379
#> 672 5 5 41.202205092 0.01989028
#> 673 5 5 41.202943764 0.01988796
#> 674 5 5 41.220596907 0.01985657
#> 675 5 5 41.300000000 0.01963030
#> 676 5 5 41.400000000 0.01939199
#> 677 5 5 41.500000000 0.01919688
#> 678 5 5 41.568563714 0.04773076
#> 679 5 5 41.600000000 0.04593829
#> 680 5 5 41.619732982 0.04486937
#> 681 5 5 41.679646228 0.04187077
#> 682 5 5 41.685514462 0.04160714
#> 683 5 5 41.700000000 0.04094204
#> 684 5 5 41.800000000 0.03684057
#> 685 5 5 41.813371757 0.03638237
#> 686 5 5 41.900000000 0.03350835
#> 687 5 5 41.946420209 0.03545269
#> 688 5 5 42.000000000 0.03371126
#> 689 5 5 42.100000000 0.03092051
#> 690 5 5 42.200000000 0.02863563
#> 691 5 5 42.300000000 0.02676493
#> 692 5 5 42.321387075 0.02641114
#> 693 5 5 42.362369453 0.02577425
#> 694 5 5 42.400000000 0.02523351
#> 695 5 5 42.425450265 0.02505416
#> 696 5 5 42.451067956 0.02475880
#> 697 5 5 42.500000000 0.02415812
#> 698 5 5 42.528349032 0.02383608
#> 699 5 5 42.600000000 0.02309906
#> 700 5 5 42.700000000 0.02223197
#> 701 5 5 42.780987439 0.03381200
#> 702 5 5 42.781909491 0.03868200
#> 703 5 5 42.800000000 0.03795830
#> 704 5 5 42.801748270 0.03808500
#> 705 5 5 42.855921660 0.03852403
#> 706 5 5 42.900000000 0.03681879
#> 707 5 5 42.993719547 0.03374613
#> 708 5 5 43.000000000 0.03355352
#> 709 5 5 43.100000000 0.03079136
#> 710 5 5 43.200000000 0.02852990
#> 711 5 5 43.299788982 0.02668191
#> 712 5 5 43.300000000 0.02667838
#> 713 5 5 43.400000000 0.02516247
#> 714 5 5 43.500000000 0.02392135
#> 715 5 5 43.535715745 0.02363063
#> 716 5 5 43.600000000 0.02298940
#> 717 5 5 43.700000000 0.02214219
#> 718 5 5 43.748596568 0.07178766
#> 719 5 5 43.800000000 0.06656351
#> 720 5 5 43.900000000 0.05781766
#> 721 5 5 44.000000000 0.05065716
#> 722 5 5 44.017009162 0.04957585
#> 723 5 5 44.100000000 0.04479496
#> 724 5 5 44.200000000 0.03999508
#> 725 5 5 44.300000000 0.03606526
#> 726 5 5 44.400000000 0.03284780
#> 727 5 5 44.498828265 0.03044731
#> 728 5 5 44.500000000 0.03041892
#> 729 5 5 44.534837082 0.22173956
#> 730 5 5 44.600000000 0.19688306
#> 731 5 5 44.603002027 0.19581594
#> 732 5 5 44.638399256 0.18369527
#> 733 5 5 44.700000000 0.16452537
#> 734 5 5 44.800000000 0.13802204
#> 735 5 5 44.809888750 0.13568403
#> 736 5 5 44.900000000 0.11632815
#> 737 5 5 45.000000000 0.09856150
#> 738 5 5 45.056258742 0.32723247
#> 739 5 5 45.079087559 0.31344550
#> 740 5 5 45.100000000 0.30135631
#> 741 5 5 45.200000000 0.25004974
#> 742 5 5 45.273440122 0.35906386
#> 743 5 5 45.287574623 0.35173706
#> 744 5 5 45.300000000 0.34355339
#> 745 5 5 45.400000000 0.28459779
#> 746 5 5 45.459455901 0.25476474
#> 747 5 5 45.500000000 0.23634828
#> 748 5 5 45.541584627 0.29005794
#> 749 5 5 45.545584023 0.28789300
#> 750 5 5 45.560062732 0.28020730
#> 751 5 5 45.566611192 0.32532263
#> 752 5 5 45.600000000 0.30549096
#> 753 5 5 45.605409941 0.32935200
#> 754 5 5 45.627068833 0.31616624
#> 755 5 5 45.636881631 0.31037774
#> 756 5 5 45.699157094 0.27617531
#> 757 5 5 45.700000000 0.27574097
#> 758 5 5 45.790076653 0.23330245
#> 759 5 5 45.792468184 0.26051922
#> 760 5 5 45.800000000 0.25689810
#> 761 5 5 45.900000000 0.21365043
#> 762 5 5 45.975007860 0.18643928
#> 763 5 5 46.000000000 0.17824230
#> 764 5 5 46.059184871 0.16038944
#> 765 5 5 46.069557597 0.15811820
#> 766 5 5 46.100000000 0.14986029
#> 767 5 5 46.174401640 0.13381473
#> 768 5 5 46.200000000 0.12805037
#> 769 5 5 46.263244569 0.11501210
#> 770 5 5 46.300000000 0.10815884
#> 771 5 5 46.301497670 0.10789013
#> 772 5 5 46.305782073 0.10712586
#> 773 5 5 46.400000000 0.09187303
#> 774 5 5 46.489457197 0.07982287
#> 775 5 5 46.500000000 0.07853953
#> 776 5 5 46.504145011 0.07804684
#> 777 5 5 46.558198689 0.07193636
#> 778 5 5 46.600000000 0.06763580
#> 779 5 5 46.696604862 0.05897070
#> 780 5 5 46.700000000 0.05869557
#> 781 5 5 46.777276693 0.05291309
#> 782 5 5 46.779176308 0.05278190
#> 783 5 5 46.800000000 0.05137595
#> 784 5 5 46.809691824 0.05074129
#> 785 5 5 46.854156565 0.04798220
#> 786 5 5 46.879281654 0.04652829
#> 787 5 5 46.900000000 0.04538314
#> 788 5 5 46.908289431 0.04493881
#> 789 5 5 46.954525982 0.04258729
#> 790 5 5 47.000000000 0.04047724
#> 791 5 5 47.100000000 0.03646002
#> 792 5 5 47.200000000 0.03317100
#> 793 5 5 47.225045373 0.03244524
#> 794 5 5 47.233253198 0.03221519
#> 795 5 5 47.300000000 0.03047820
#> 796 5 5 47.316137190 0.03009194
#> 797 5 5 47.387794884 0.02851957
#> 798 5 5 47.400000000 0.02827351
#> 799 5 5 47.500000000 0.02646845
#> 800 5 5 47.600000000 0.02499060
#> 801 5 5 47.700000000 0.02378063
#> 802 5 5 47.734537359 0.02341588
#> 803 5 5 47.800000000 0.02279000
#> 804 5 5 47.900000000 0.02197893
#> 805 5 5 48.000000000 0.02131489
#> 806 5 5 48.006842608 0.02127413
#> 807 5 5 48.100000000 0.02077122
#> 808 5 5 48.200000000 0.02032610
#> 809 5 5 48.300000000 0.01996166
#> 810 5 5 48.302227992 0.01995435
#> 811 5 5 48.400000000 0.01966329
#> 812 5 5 48.439774218 0.03712440
#> 813 5 5 48.446914691 0.03685770
#> 814 5 5 48.500000000 0.03498998
#> 815 5 5 48.507946620 0.03472707
#> 816 5 5 48.600000000 0.03196744
#> 817 5 5 48.624882588 0.03132252
#> 818 5 5 48.638839838 0.03096471
#> 819 5 5 48.669928147 0.03020250
#> 820 5 5 48.700000000 0.02950866
#> 821 5 5 48.800000000 0.02747971
#> 822 5 5 48.806357041 0.02736393
#> 823 5 5 48.816478817 0.02718260
#> 824 5 5 48.858734446 0.02646404
#> 825 5 5 48.900000000 0.02581854
#> 826 5 5 48.930369694 0.02537640
#> 827 5 5 49.000000000 0.02445851
#> 828 5 5 49.100000000 0.02334500
#> 829 5 5 49.200000000 0.02243333
#> 830 5 5 49.300000000 0.02168692
#> 831 5 5 49.382912433 0.02117177
#> 832 5 5 49.400000000 0.02107581
#> 833 5 5 49.500000000 0.02057548
#> 834 5 5 49.600000000 0.02016584
#> 835 5 5 49.700000000 0.01983045
#> 836 5 5 49.800000000 0.01955586
#> 837 5 5 49.900000000 0.01933105
#> 838 5 5 50.000000000 0.01914699