Compute the conditional intensity at a specified time

Compute the conditional intensity at a specified time

Usage

spatial_conditional_intensity(hawkes, parameters, time, stepsize)

Arguments

hawkes

A hawkes object.

parameters

Named list containing background, triggering, spatial, and temporal parameters.

time

Time point where the conditional intensity is evaluated.

stepsize

Grid cell size used to evaluate the spatial intensity.

Value

A numeric vector.

Examples

spatial_region <- create_rectangular_sf(0, 10, 0, 10)

params <- list(
  background_rate = list(intercept = -4),
  triggering_rate = 0.75,
  spatial = list(mean = 0, sd = 0.75),
  temporal = list(rate = 2)
)
hawkes <- rHawkes(params, time_window = c(0, 50), spatial_region = spatial_region)
spatial_conditional_intensity(hawkes, params, 25, 0.5)
#>        x    y  t  intensity
#> 1   0.25 0.25 25 0.01905261
#> 2   0.75 0.25 25 0.02024906
#> 3   1.25 0.25 25 0.02179254
#> 4   1.75 0.25 25 0.02285835
#> 5   2.25 0.25 25 0.02321678
#> 6   2.75 0.25 25 0.02415044
#> 7   3.25 0.25 25 0.02769186
#> 8   3.75 0.25 25 0.03321855
#> 9   4.25 0.25 25 0.03569850
#> 10  4.75 0.25 25 0.03189081
#> 11  5.25 0.25 25 0.02532983
#> 12  5.75 0.25 25 0.02074961
#> 13  6.25 0.25 25 0.01890053
#> 14  6.75 0.25 25 0.01841693
#> 15  7.25 0.25 25 0.01832876
#> 16  7.75 0.25 25 0.01831695
#> 17  8.25 0.25 25 0.01831574
#> 18  8.75 0.25 25 0.01831565
#> 19  9.25 0.25 25 0.01831564
#> 20  9.75 0.25 25 0.01831564
#> 21  0.25 0.75 25 0.01894923
#> 22  0.75 0.75 25 0.01998453
#> 23  1.25 0.75 25 0.02166359
#> 24  1.75 0.75 25 0.02376354
#> 25  2.25 0.75 25 0.02668389
#> 26  2.75 0.75 25 0.03361569
#> 27  3.25 0.75 25 0.05007020
#> 28  3.75 0.75 25 0.07305763
#> 29  4.25 0.75 25 0.08415236
#> 30  4.75 0.75 25 0.07117115
#> 31  5.25 0.75 25 0.04673823
#> 32  5.75 0.75 25 0.02878281
#> 33  6.25 0.75 25 0.02103443
#> 34  6.75 0.75 25 0.01882527
#> 35  7.25 0.75 25 0.01838491
#> 36  7.75 0.75 25 0.01832242
#> 37  8.25 0.75 25 0.01831612
#> 38  8.75 0.75 25 0.01831567
#> 39  9.25 0.75 25 0.01831564
#> 40  9.75 0.75 25 0.01831564
#> 41  0.25 1.25 25 0.01939218
#> 42  0.75 1.25 25 0.02001406
#> 43  1.25 1.25 25 0.02133903
#> 44  1.75 1.25 25 0.02426204
#> 45  2.25 1.25 25 0.03095836
#> 46  2.75 1.25 25 0.04937789
#> 47  3.25 1.25 25 0.09259065
#> 48  3.75 1.25 25 0.15359765
#> 49  4.25 1.25 25 0.18764568
#> 50  4.75 1.25 25 0.16175025
#> 51  5.25 1.25 25 0.10196356
#> 52  5.75 1.25 25 0.05279333
#> 53  6.25 1.25 25 0.02853770
#> 54  6.75 1.25 25 0.02048796
#> 55  7.25 1.25 25 0.01863964
#> 56  7.75 1.25 25 0.01834879
#> 57  8.25 1.25 25 0.01831797
#> 58  8.75 1.25 25 0.01831579
#> 59  9.25 1.25 25 0.01831571
#> 60  9.75 1.25 25 0.01831573
#> 61  0.25 1.75 25 0.02463799
#> 62  0.75 1.75 25 0.02425254
#> 63  1.25 1.75 25 0.02321340
#> 64  1.75 1.75 25 0.02474951
#> 65  2.25 1.75 25 0.03359293
#> 66  2.75 1.75 25 0.06341233
#> 67  3.25 1.75 25 0.13699336
#> 68  3.75 1.75 25 0.24706226
#> 69  4.25 1.75 25 0.32308959
#> 70  4.75 1.75 25 0.30027314
#> 71  5.25 1.75 25 0.20334285
#> 72  5.75 1.75 25 0.10559515
#> 73  6.25 1.75 25 0.04770123
#> 74  6.75 1.75 25 0.02521713
#> 75  7.25 1.75 25 0.01941816
#> 76  7.75 1.75 25 0.01843352
#> 77  8.25 1.75 25 0.01832439
#> 78  8.75 1.75 25 0.01831673
#> 79  9.25 1.75 25 0.01831683
#> 80  9.75 1.75 25 0.01831723
#> 81  0.25 2.25 25 0.05121812
#> 82  0.75 2.25 25 0.04687520
#> 83  1.25 2.25 25 0.03515100
#> 84  1.75 2.25 25 0.02781629
#> 85  2.25 2.25 25 0.03280366
#> 86  2.75 2.25 25 0.06427190
#> 87  3.25 2.25 25 0.14975582
#> 88  3.75 2.25 25 0.29143135
#> 89  4.25 2.25 25 0.41818032
#> 90  4.75 2.25 25 0.43526605
#> 91  5.25 2.25 25 0.33050767
#> 92  5.75 2.25 25 0.18457340
#> 93  6.25 2.25 25 0.07986962
#> 94  6.75 2.25 25 0.03378378
#> 95  7.25 2.25 25 0.02090915
#> 96  7.75 2.25 25 0.01860549
#> 97  8.25 2.25 25 0.01834114
#> 98  8.75 2.25 25 0.01832409
#> 99  9.25 2.25 25 0.01832850
#> 100 9.75 2.25 25 0.01833304
#> 101 0.25 2.75 25 0.13136253
#> 102 0.75 2.75 25 0.11529785
#> 103 1.25 2.75 25 0.07227322
#> 104 1.75 2.75 25 0.03946312
#> 105 2.25 2.75 25 0.03112224
#> 106 2.75 2.75 25 0.05189545
#> 107 3.25 2.75 25 0.12234435
#> 108 3.75 2.75 25 0.25760272
#> 109 4.25 2.75 25 0.41346626
#> 110 4.75 2.75 25 0.48644613
#> 111 5.25 2.75 25 0.41153970
#> 112 5.75 2.75 25 0.24742416
#> 113 6.25 2.75 25 0.10887477
#> 114 6.75 2.75 25 0.04221081
#> 115 7.25 2.75 25 0.02249363
#> 116 7.75 2.75 25 0.01881443
#> 117 8.25 2.75 25 0.01838219
#> 118 8.75 2.75 25 0.01836817
#> 119 9.25 2.75 25 0.01840521
#> 120 9.75 2.75 25 0.01843777
#> 121 0.25 3.25 25 0.26942638
#> 122 0.75 3.25 25 0.23292290
#> 123 1.25 3.25 25 0.13636175
#> 124 1.75 3.25 25 0.06095905
#> 125 2.25 3.25 25 0.03190874
#> 126 2.75 3.25 25 0.03704217
#> 127 3.25 3.25 25 0.07930600
#> 128 3.75 3.25 25 0.17634416
#> 129 4.25 3.25 25 0.31338943
#> 130 4.75 3.25 25 0.40809648
#> 131 5.25 3.25 25 0.37499382
#> 132 5.75 3.25 25 0.23987461
#> 133 6.25 3.25 25 0.11035698
#> 134 6.75 3.25 25 0.04367061
#> 135 7.25 3.25 25 0.02296472
#> 136 7.75 3.25 25 0.01895173
#> 137 8.25 3.25 25 0.01848788
#> 138 8.75 3.25 25 0.01854073
#> 139 9.25 3.25 25 0.01871680
#> 140 9.75 3.25 25 0.01886555
#> 141 0.25 3.75 25 0.37875288
#> 142 0.75 3.75 25 0.32538999
#> 143 1.25 3.75 25 0.18644445
#> 144 1.75 3.75 25 0.07787628
#> 145 2.25 3.75 25 0.03332057
#> 146 2.75 3.75 25 0.02717854
#> 147 3.25 3.75 25 0.04541307
#> 148 3.75 3.75 25 0.09682553
#> 149 4.25 3.75 25 0.18079363
#> 150 4.75 3.75 25 0.25129161
#> 151 5.25 3.75 25 0.24545144
#> 152 5.75 3.75 25 0.16681611
#> 153 6.25 3.75 25 0.08286338
#> 154 6.75 3.75 25 0.03692246
#> 155 7.25 3.75 25 0.02197580
#> 156 7.75 3.75 25 0.01899523
#> 157 8.25 3.75 25 0.01870901
#> 158 8.75 3.75 25 0.01895032
#> 159 9.25 3.75 25 0.01947066
#> 160 9.75 3.75 25 0.01990428
#> 161 0.25 4.25 25 0.35376458
#> 162 0.75 4.25 25 0.30290427
#> 163 1.25 4.25 25 0.17347035
#> 164 1.75 4.25 25 0.07283983
#> 165 2.25 4.25 25 0.03109754
#> 166 2.75 4.25 25 0.02221217
#> 167 3.25 4.25 25 0.02740927
#> 168 3.75 4.25 25 0.04693651
#> 169 4.25 4.25 25 0.08238290
#> 170 4.75 4.25 25 0.11616756
#> 171 5.25 4.25 25 0.11889719
#> 172 5.75 4.25 25 0.08728868
#> 173 6.25 4.25 25 0.04973297
#> 174 6.75 4.25 25 0.02790808
#> 175 7.25 4.25 25 0.02050514
#> 176 7.75 4.25 25 0.01902938
#> 177 8.25 4.25 25 0.01899745
#> 178 8.75 4.25 25 0.01948401
#> 179 9.25 4.25 25 0.02045539
#> 180 9.75 4.25 25 0.02126092
#> 181 0.25 4.75 25 0.22201773
#> 182 0.75 4.75 25 0.19008102
#> 183 1.25 4.75 25 0.11147968
#> 184 1.75 4.75 25 0.05089298
#> 185 2.25 4.75 25 0.02578335
#> 186 2.75 4.75 25 0.01987885
#> 187 3.25 4.75 25 0.02058335
#> 188 3.75 4.75 25 0.02577417
#> 189 4.25 4.75 25 0.03608696
#> 190 4.75 4.75 25 0.04699455
#> 191 5.25 4.75 25 0.04933266
#> 192 5.75 4.75 25 0.04068638
#> 193 6.25 4.75 25 0.02914046
#> 194 6.75 4.75 25 0.02205835
#> 195 7.25 4.75 25 0.01959967
#> 196 7.75 4.75 25 0.01911433
#> 197 8.25 4.75 25 0.01917890
#> 198 8.75 4.75 25 0.01973141
#> 199 9.25 4.75 25 0.02087207
#> 200 9.75 4.75 25 0.02182192
#> 201 0.25 5.25 25 0.09987723
#> 202 0.75 5.25 25 0.08661968
#> 203 1.25 5.25 25 0.05531468
#> 204 1.75 5.25 25 0.03139651
#> 205 2.25 5.25 25 0.02141800
#> 206 2.75 5.25 25 0.01890475
#> 207 3.25 5.25 25 0.01875450
#> 208 3.75 5.25 25 0.01969440
#> 209 4.25 5.25 25 0.02179089
#> 210 4.75 5.25 25 0.02427242
#> 211 5.25 5.25 25 0.02519690
#> 212 5.75 5.25 25 0.02372954
#> 213 6.25 5.25 25 0.02142020
#> 214 6.75 5.25 25 0.01997780
#> 215 7.25 5.25 25 0.01947511
#> 216 7.75 5.25 25 0.01926762
#> 217 8.25 5.25 25 0.01917401
#> 218 8.75 5.25 25 0.01947877
#> 219 9.25 5.25 25 0.02029645
#> 220 9.75 5.25 25 0.02099913
#> 221 0.25 5.75 25 0.04036756
#> 222 0.75 5.75 25 0.03700280
#> 223 1.25 5.75 25 0.02899856
#> 224 1.75 5.75 25 0.02268700
#> 225 2.25 5.75 25 0.01971382
#> 226 2.75 5.75 25 0.01869201
#> 227 3.25 5.75 25 0.01844059
#> 228 3.75 5.75 25 0.01851905
#> 229 4.25 5.75 25 0.01885641
#> 230 4.75 5.75 25 0.01938418
#> 231 5.25 5.75 25 0.01982248
#> 232 5.75 5.75 25 0.01992177
#> 233 6.25 5.75 25 0.01983317
#> 234 6.75 5.75 25 0.01982670
#> 235 7.25 5.75 25 0.01977902
#> 236 7.75 5.75 25 0.01947011
#> 237 8.25 5.75 25 0.01909249
#> 238 8.75 5.75 25 0.01901969
#> 239 9.25 5.75 25 0.01932743
#> 240 9.75 5.75 25 0.01964012
#> 241 0.25 6.25 25 0.02280668
#> 242 0.75 6.25 25 0.02281797
#> 243 1.25 6.25 25 0.02193878
#> 244 1.75 6.25 25 0.02073498
#> 245 2.25 6.25 25 0.01956941
#> 246 2.75 6.25 25 0.01877925
#> 247 3.25 6.25 25 0.01843950
#> 248 3.75 6.25 25 0.01837196
#> 249 4.25 6.25 25 0.01845518
#> 250 4.75 6.25 25 0.01869523
#> 251 5.25 6.25 25 0.01906566
#> 252 5.75 6.25 25 0.01943429
#> 253 6.25 6.25 25 0.01974736
#> 254 6.75 6.25 25 0.02001023
#> 255 7.25 6.25 25 0.02000252
#> 256 7.75 6.25 25 0.01956447
#> 257 8.25 6.25 25 0.01899962
#> 258 8.75 6.25 25 0.01869089
#> 259 9.25 6.25 25 0.01867631
#> 260 9.75 6.25 25 0.01874255
#> 261 0.25 6.75 25 0.01943078
#> 262 0.75 6.75 25 0.02017039
#> 263 1.25 6.75 25 0.02072463
#> 264 1.75 6.75 25 0.02054707
#> 265 2.25 6.75 25 0.01972956
#> 266 2.75 6.75 25 0.01892868
#> 267 3.25 6.75 25 0.01850466
#> 268 3.75 6.75 25 0.01837356
#> 269 4.25 6.75 25 0.01839867
#> 270 4.75 6.75 25 0.01855747
#> 271 5.25 6.75 25 0.01884271
#> 272 5.75 6.75 25 0.01917699
#> 273 6.25 6.75 25 0.01951089
#> 274 6.75 6.75 25 0.01978948
#> 275 7.25 6.75 25 0.01977809
#> 276 7.75 6.75 25 0.01936217
#> 277 8.25 6.75 25 0.01883830
#> 278 8.75 6.75 25 0.01852216
#> 279 9.25 6.75 25 0.01842491
#> 280 9.75 6.75 25 0.01841260
#> 281 0.25 7.25 25 0.01888640
#> 282 0.75 7.25 25 0.01968998
#> 283 1.25 7.25 25 0.02063502
#> 284 1.75 7.25 25 0.02106770
#> 285 2.25 7.25 25 0.02064802
#> 286 2.75 7.25 25 0.01973600
#> 287 3.25 7.25 25 0.01893026
#> 288 3.75 7.25 25 0.01850755
#> 289 4.25 7.25 25 0.01839235
#> 290 4.75 7.25 25 0.01844598
#> 291 5.25 7.25 25 0.01859496
#> 292 5.75 7.25 25 0.01879287
#> 293 6.25 7.25 25 0.01901858
#> 294 6.75 7.25 25 0.01921319
#> 295 7.25 7.25 25 0.01920749
#> 296 7.75 7.25 25 0.01894333
#> 297 8.25 7.25 25 0.01862003
#> 298 8.75 7.25 25 0.01842567
#> 299 9.25 7.25 25 0.01835727
#> 300 9.75 7.25 25 0.01833907
#> 301 0.25 7.75 25 0.01888772
#> 302 0.75 7.75 25 0.02016082
#> 303 1.25 7.75 25 0.02251278
#> 304 1.75 7.75 25 0.02500653
#> 305 2.25 7.75 25 0.02566991
#> 306 2.75 7.75 25 0.02377837
#> 307 3.25 7.75 25 0.02101067
#> 308 3.75 7.75 25 0.01919229
#> 309 4.25 7.75 25 0.01851957
#> 310 4.75 7.75 25 0.01839617
#> 311 5.25 7.75 25 0.01843092
#> 312 5.75 7.75 25 0.01850616
#> 313 6.25 7.75 25 0.01860473
#> 314 6.75 7.75 25 0.01869519
#> 315 7.25 7.75 25 0.01869638
#> 316 7.75 7.75 25 0.01858487
#> 317 8.25 7.75 25 0.01844911
#> 318 8.75 7.75 25 0.01836774
#> 319 9.25 7.75 25 0.01833734
#> 320 9.75 7.75 25 0.01832647
#> 321 0.25 8.25 25 0.01928611
#> 322 0.75 8.25 25 0.02212271
#> 323 1.25 8.25 25 0.02850946
#> 324 1.75 8.25 25 0.03667382
#> 325 2.25 8.25 25 0.04023661
#> 326 2.75 8.25 25 0.03547406
#> 327 3.25 8.25 25 0.02705077
#> 328 3.75 8.25 25 0.02119898
#> 329 4.25 8.25 25 0.01894623
#> 330 4.75 8.25 25 0.01843614
#> 331 5.25 8.25 25 0.01837686
#> 332 5.75 8.25 25 0.01838582
#> 333 6.25 8.25 25 0.01840923
#> 334 6.75 8.25 25 0.01843811
#> 335 7.25 8.25 25 0.01844462
#> 336 7.75 8.25 25 0.01841458
#> 337 8.25 8.25 25 0.01837144
#> 338 8.75 8.25 25 0.01834174
#> 339 9.25 8.25 25 0.01832804
#> 340 9.75 8.25 25 0.01832169
#> 341 0.25 8.75 25 0.01989715
#> 342 0.75 8.75 25 0.02513048
#> 343 1.25 8.75 25 0.03786554
#> 344 1.75 8.75 25 0.05521699
#> 345 2.25 8.75 25 0.06373684
#> 346 2.75 8.75 25 0.05454861
#> 347 3.25 8.75 25 0.03697503
#> 348 3.75 8.75 25 0.02451280
#> 349 4.25 8.75 25 0.01965922
#> 350 4.75 8.75 25 0.01853267
#> 351 5.25 8.75 25 0.01837128
#> 352 5.75 8.75 25 0.01835208
#> 353 6.25 8.75 25 0.01835001
#> 354 6.75 8.75 25 0.01835903
#> 355 7.25 8.75 25 0.01836950
#> 356 7.75 8.75 25 0.01836771
#> 357 8.25 8.75 25 0.01835303
#> 358 8.75 8.75 25 0.01833603
#> 359 9.25 8.75 25 0.01832460
#> 360 9.75 8.75 25 0.01831903
#> 361 0.25 9.25 25 0.02022278
#> 362 0.75 9.25 25 0.02696647
#> 363 1.25 9.25 25 0.04401144
#> 364 1.75 9.25 25 0.06792258
#> 365 2.25 9.25 25 0.08024527
#> 366 2.75 9.25 25 0.06814927
#> 367 3.25 9.25 25 0.04411478
#> 368 3.75 9.25 25 0.02690870
#> 369 4.25 9.25 25 0.02017359
#> 370 4.75 9.25 25 0.01859862
#> 371 5.25 9.25 25 0.01836683
#> 372 5.75 9.25 25 0.01833823
#> 373 6.25 9.25 25 0.01833289
#> 374 6.75 9.25 25 0.01833918
#> 375 7.25 9.25 25 0.01835277
#> 376 7.75 9.25 25 0.01836131
#> 377 8.25 9.25 25 0.01835558
#> 378 8.75 9.25 25 0.01833989
#> 379 9.25 9.25 25 0.01832580
#> 380 9.75 9.25 25 0.01831863
#> 381 0.25 9.75 25 0.01989758
#> 382 0.75 9.75 25 0.02569525
#> 383 1.25 9.75 25 0.04063238
#> 384 1.75 9.75 25 0.06188426
#> 385 2.25 9.75 25 0.07308217
#> 386 2.75 9.75 25 0.06256976
#> 387 3.25 9.75 25 0.04128393
#> 388 3.75 9.75 25 0.02597636
#> 389 4.25 9.75 25 0.01997006
#> 390 4.75 9.75 25 0.01856091
#> 391 5.25 9.75 25 0.01835181
#> 392 5.75 9.75 25 0.01832726
#> 393 6.25 9.75 25 0.01832435
#> 394 6.75 9.75 25 0.01833100
#> 395 7.25 9.75 25 0.01834600
#> 396 7.75 9.75 25 0.01835963
#> 397 8.25 9.75 25 0.01835889
#> 398 8.75 9.75 25 0.01834385
#> 399 9.25 9.75 25 0.01832772
#> 400 9.75 9.75 25 0.01831902

params <- list(
  background_rate = list(intercept = -4.5, z = 1),
  triggering_rate = 0.5,
  spatial = list(mean = 0, sd = 0.75),
  temporal = list(rate = 2),
  fixed = list(spatial = "mean")
)
data("example_background_covariates")
hawkes <- rHawkes(
  params,
  c(0, 50),
  example_background_covariates,
  covariate_columns = "z",
  spatial_burnin = 1
)

spatial_conditional_intensity(hawkes, params, 25, 0.5)
#>        x    y  t   intensity
#> 1   0.25 0.25 25 0.006418063
#> 2   0.75 0.25 25 0.006421499
#> 3   0.25 0.75 25 0.006506497
#> 4   0.75 0.75 25 0.006560684
#> 5   1.25 0.25 25 0.006424471
#> 6   1.75 0.25 25 0.006421158
#> 7   1.25 0.75 25 0.006564998
#> 8   1.75 0.75 25 0.006512792
#> 9   2.25 0.25 25 0.006430296
#> 10  2.25 0.75 25 0.006472867
#> 11  2.75 0.25 25 0.006461761
#> 12  3.25 0.25 25 0.006463696
#> 13  2.75 0.75 25 0.006492216
#> 14  3.25 0.75 25 0.006521211
#> 15  3.75 0.25 25 0.006517134
#> 16  4.25 0.25 25 0.006579611
#> 17  3.75 0.75 25 0.006624211
#> 18  4.25 0.75 25 0.006822852
#> 19  4.75 0.25 25 0.007087348
#> 20  5.25 0.25 25 0.009461336
#> 21  4.75 0.75 25 0.008279043
#> 22  5.25 0.75 25 0.015385217
#> 23  5.75 0.25 25 0.016459309
#> 24  6.25 0.25 25 0.028187683
#> 25  5.75 0.75 25 0.036442658
#> 26  6.25 0.75 25 0.071657420
#> 27  6.75 0.25 25 0.036746021
#> 28  7.25 0.25 25 0.033585993
#> 29  6.75 0.75 25 0.097427260
#> 30  7.25 0.75 25 0.087914653
#> 31  7.75 0.25 25 0.022005834
#> 32  7.75 0.75 25 0.053198672
#> 33  8.25 0.25 25 0.012151399
#> 34  8.75 0.25 25 0.007769356
#> 35  8.25 0.75 25 0.023635736
#> 36  8.75 0.75 25 0.010480589
#> 37  9.25 0.25 25 0.006617114
#> 38  9.75 0.25 25 0.006431979
#> 39  9.25 0.75 25 0.007027535
#> 40  9.75 0.75 25 0.006471816
#> 41  0.25 1.25 25 0.007369801
#> 42  0.75 1.25 25 0.007917800
#> 43  0.25 1.75 25 0.012606270
#> 44  0.75 1.75 25 0.016160040
#> 45  1.25 1.25 25 0.007953759
#> 46  1.75 1.25 25 0.007426413
#> 47  1.25 1.75 25 0.016277989
#> 48  1.75 1.75 25 0.012864404
#> 49  2.25 1.25 25 0.006992140
#> 50  2.25 1.75 25 0.009559035
#> 51  2.75 1.25 25 0.007115824
#> 52  3.25 1.25 25 0.007421311
#> 53  2.75 1.75 25 0.009073884
#> 54  3.25 1.75 25 0.011187215
#> 55  3.75 1.25 25 0.008263283
#> 56  4.25 1.25 25 0.008764484
#> 57  3.75 1.75 25 0.014668955
#> 58  4.25 1.75 25 0.016120601
#> 59  4.75 1.25 25 0.011498689
#> 60  5.25 1.25 25 0.024967587
#> 61  4.75 1.75 25 0.017982912
#> 62  5.25 1.75 25 0.033162253
#> 63  5.75 1.25 25 0.065402424
#> 64  6.25 1.25 25 0.133303520
#> 65  5.75 1.75 25 0.083427622
#> 66  6.25 1.75 25 0.168043268
#> 67  6.75 1.25 25 0.182523941
#> 68  7.25 1.25 25 0.163926757
#> 69  6.75 1.75 25 0.228201257
#> 70  7.25 1.75 25 0.203318725
#> 71  7.75 1.25 25 0.096685509
#> 72  7.75 1.75 25 0.118759928
#> 73  8.25 1.25 25 0.039612516
#> 74  8.75 1.25 25 0.014266167
#> 75  8.25 1.75 25 0.047614836
#> 76  8.75 1.75 25 0.016138156
#> 77  9.25 1.25 25 0.007600049
#> 78  9.75 1.25 25 0.006529991
#> 79  9.25 1.75 25 0.007882389
#> 80  9.75 1.75 25 0.006557483
#> 81  0.25 2.25 25 0.032168914
#> 82  0.75 2.25 25 0.046949432
#> 83  1.25 2.25 25 0.047469685
#> 84  1.75 2.25 25 0.033318950
#> 85  2.25 2.25 25 0.019771867
#> 86  2.75 2.25 25 0.018436339
#> 87  3.25 2.25 25 0.028001287
#> 88  3.75 2.25 25 0.044190004
#> 89  4.25 2.25 25 0.049275794
#> 90  4.75 2.25 25 0.045707359
#> 91  5.25 2.25 25 0.051148667
#> 92  5.75 2.25 25 0.090528350
#> 93  6.25 2.25 25 0.160656270
#> 94  6.75 2.25 25 0.203183726
#> 95  7.25 2.25 25 0.175162910
#> 96  7.75 2.25 25 0.100050548
#> 97  8.25 2.25 25 0.040174714
#> 98  8.75 2.25 25 0.014392202
#> 99  9.25 2.25 25 0.007614800
#> 100 9.75 2.25 25 0.006543672
#> 101 0.25 2.75 25 0.075090541
#> 102 0.75 2.75 25 0.114533147
#> 103 0.25 3.25 25 0.123823215
#> 104 0.75 3.25 25 0.191434620
#> 105 1.25 2.75 25 0.116072249
#> 106 1.75 2.75 25 0.078569976
#> 107 1.25 3.25 25 0.194311861
#> 108 1.75 3.25 25 0.131063987
#> 109 2.25 2.75 25 0.043336313
#> 110 2.25 3.25 25 0.071650386
#> 111 2.75 2.75 25 0.043661949
#> 112 3.25 2.75 25 0.072677658
#> 113 2.75 3.25 25 0.073630227
#> 114 3.25 3.25 25 0.136148626
#> 115 3.75 2.75 25 0.130289492
#> 116 4.25 2.75 25 0.146190693
#> 117 3.75 3.25 25 0.241849285
#> 118 4.25 3.25 25 0.282195419
#> 119 4.75 2.75 25 0.140422008
#> 120 5.25 2.75 25 0.122919348
#> 121 4.75 3.25 25 0.259348416
#> 122 5.25 3.25 25 0.208270187
#> 123 5.75 2.75 25 0.129138113
#> 124 6.25 2.75 25 0.171411418
#> 125 5.75 3.25 25 0.188696195
#> 126 6.25 3.25 25 0.206170522
#> 127 6.75 2.75 25 0.168782837
#> 128 7.25 2.75 25 0.130979539
#> 129 6.75 3.25 25 0.172840707
#> 130 7.25 3.25 25 0.114824252
#> 131 7.75 2.75 25 0.067782952
#> 132 7.75 3.25 25 0.052004161
#> 133 8.25 2.75 25 0.026990803
#> 134 8.75 2.75 25 0.011354665
#> 135 8.25 3.25 25 0.019873097
#> 136 8.75 3.25 25 0.009586576
#> 137 9.25 2.75 25 0.007155954
#> 138 9.75 2.75 25 0.006532771
#> 139 9.25 3.25 25 0.006920379
#> 140 9.75 3.25 25 0.006521715
#> 141 0.25 3.75 25 0.135310800
#> 142 0.75 3.75 25 0.209992626
#> 143 0.25 4.25 25 0.097376246
#> 144 0.75 4.25 25 0.150979085
#> 145 1.25 3.75 25 0.214797255
#> 146 1.75 3.75 25 0.147803605
#> 147 1.25 4.25 25 0.157127103
#> 148 1.75 4.25 25 0.114367896
#> 149 2.25 3.75 25 0.089537782
#> 150 2.25 4.25 25 0.083930176
#> 151 2.75 3.75 25 0.119944705
#> 152 3.25 3.75 25 0.224825832
#> 153 2.75 4.25 25 0.140721791
#> 154 3.25 4.25 25 0.286910743
#> 155 3.75 3.75 25 0.477966479
#> 156 4.25 3.75 25 0.562790040
#> 157 3.75 4.25 25 0.603758108
#> 158 4.25 4.25 25 0.746552251
#> 159 4.75 3.75 25 0.667247986
#> 160 5.25 3.75 25 0.575471293
#> 161 4.75 4.25 25 0.861902788
#> 162 5.25 4.25 25 0.729775177
#> 163 5.75 3.75 25 0.380829258
#> 164 6.25 3.75 25 0.366314038
#> 165 5.75 4.25 25 0.478102631
#> 166 6.25 4.25 25 0.420815240
#> 167 6.75 3.75 25 0.221576443
#> 168 7.25 3.75 25 0.142887630
#> 169 6.75 4.25 25 0.251766720
#> 170 7.25 4.25 25 0.158443477
#> 171 7.75 3.75 25 0.056373792
#> 172 7.75 4.25 25 0.062730411
#> 173 8.25 3.75 25 0.020127916
#> 174 8.75 3.75 25 0.009934013
#> 175 8.25 4.25 25 0.022032294
#> 176 8.75 4.25 25 0.010367520
#> 177 9.25 3.75 25 0.007016304
#> 178 9.75 3.75 25 0.006588464
#> 179 9.25 4.25 25 0.007099839
#> 180 9.75 4.25 25 0.006602733
#> 181 0.25 4.75 25 0.047920705
#> 182 0.75 4.75 25 0.073461478
#> 183 0.25 5.25 25 0.018844449
#> 184 0.75 5.25 25 0.027326451
#> 185 1.25 4.75 25 0.079931949
#> 186 1.75 4.75 25 0.066043347
#> 187 1.25 5.25 25 0.032307376
#> 188 1.75 5.25 25 0.032742151
#> 189 2.25 4.75 25 0.067856653
#> 190 2.25 5.25 25 0.047571829
#> 191 2.75 4.75 25 0.156599989
#> 192 3.25 4.75 25 0.320150704
#> 193 2.75 5.25 25 0.131780555
#> 194 3.25 5.25 25 0.266312349
#> 195 3.75 4.75 25 0.798079232
#> 196 4.25 4.75 25 0.975299770
#> 197 3.75 5.25 25 0.701510163
#> 198 4.25 5.25 25 0.853496169
#> 199 4.75 4.75 25 1.391373185
#> 200 5.25 4.75 25 1.243387242
#> 201 4.75 5.25 25 1.280654942
#> 202 5.25 5.25 25 1.162760215
#> 203 5.75 4.75 25 0.674622809
#> 204 6.25 4.75 25 0.590755011
#> 205 5.75 5.25 25 0.617280243
#> 206 6.25 5.25 25 0.544632049
#> 207 6.75 4.75 25 0.275395518
#> 208 7.25 4.75 25 0.176365037
#> 209 6.75 5.25 25 0.238984263
#> 210 7.25 5.25 25 0.153769737
#> 211 7.75 4.75 25 0.065391081
#> 212 7.75 5.25 25 0.055361099
#> 213 8.25 4.75 25 0.022372263
#> 214 8.75 4.75 25 0.010462865
#> 215 8.25 5.25 25 0.019141679
#> 216 8.75 5.25 25 0.009638481
#> 217 9.25 4.75 25 0.007062622
#> 218 9.75 4.75 25 0.006609718
#> 219 9.25 5.25 25 0.006880624
#> 220 9.75 5.25 25 0.006575820
#> 221 0.25 5.75 25 0.008969138
#> 222 0.75 5.75 25 0.011103403
#> 223 0.25 6.25 25 0.006857854
#> 224 0.75 6.25 25 0.007296089
#> 225 1.25 5.75 25 0.013824420
#> 226 1.75 5.75 25 0.016343499
#> 227 1.25 6.25 25 0.008619059
#> 228 1.75 6.25 25 0.009809478
#> 229 2.25 5.75 25 0.028218399
#> 230 2.25 6.25 25 0.016938721
#> 231 2.75 5.75 25 0.078672240
#> 232 3.25 5.75 25 0.155194010
#> 233 2.75 6.25 25 0.048741155
#> 234 3.25 6.25 25 0.078460692
#> 235 3.75 5.75 25 0.370527123
#> 236 4.25 5.75 25 0.458124781
#> 237 3.75 6.25 25 0.223723764
#> 238 4.25 6.25 25 0.258376566
#> 239 4.75 5.75 25 0.605267495
#> 240 5.25 5.75 25 0.541973878
#> 241 4.75 6.25 25 0.406622256
#> 242 5.25 6.25 25 0.383501646
#> 243 5.75 5.75 25 0.334048276
#> 244 6.25 5.75 25 0.293683100
#> 245 5.75 6.25 25 0.214160050
#> 246 6.25 6.25 25 0.198614063
#> 247 6.75 5.75 25 0.148602790
#> 248 7.25 5.75 25 0.095711629
#> 249 6.75 6.25 25 0.079260092
#> 250 7.25 6.25 25 0.056957599
#> 251 7.75 5.75 25 0.035962284
#> 252 7.75 6.25 25 0.020799341
#> 253 8.25 5.75 25 0.013812479
#> 254 8.75 5.75 25 0.008370000
#> 255 8.25 6.25 25 0.009743762
#> 256 8.75 6.25 25 0.007602651
#> 257 9.25 5.75 25 0.006674830
#> 258 9.75 5.75 25 0.006525030
#> 259 9.25 6.25 25 0.006565116
#> 260 9.75 6.25 25 0.006511722
#> 261 0.25 6.75 25 0.006499063
#> 262 0.75 6.75 25 0.006571454
#> 263 0.25 7.25 25 0.006463361
#> 264 0.75 7.25 25 0.006472218
#> 265 1.25 6.75 25 0.007036620
#> 266 1.75 6.75 25 0.007348687
#> 267 1.25 7.25 25 0.006809788
#> 268 1.75 7.25 25 0.006865147
#> 269 2.25 6.75 25 0.009910523
#> 270 2.25 7.25 25 0.008635073
#> 271 2.75 6.75 25 0.019232978
#> 272 3.25 6.75 25 0.027669616
#> 273 2.75 7.25 25 0.015328332
#> 274 3.25 7.25 25 0.017594301
#> 275 3.75 6.75 25 0.056742526
#> 276 4.25 6.75 25 0.067904165
#> 277 3.75 7.25 25 0.037743118
#> 278 4.25 7.25 25 0.042338491
#> 279 4.75 6.75 25 0.083835709
#> 280 5.25 6.75 25 0.077170237
#> 281 4.75 7.25 25 0.058430604
#> 282 5.25 7.25 25 0.056283832
#> 283 5.75 6.75 25 0.055273422
#> 284 6.25 6.75 25 0.049678838
#> 285 5.75 7.25 25 0.038141804
#> 286 6.25 7.25 25 0.035001931
#> 287 6.75 6.75 25 0.027902856
#> 288 7.25 6.75 25 0.021519110
#> 289 6.75 7.25 25 0.016995397
#> 290 7.25 7.25 25 0.015569139
#> 291 7.75 6.75 25 0.011050920
#> 292 7.75 7.25 25 0.008822775
#> 293 8.25 6.75 25 0.007465953
#> 294 8.75 6.75 25 0.006896926
#> 295 8.25 7.25 25 0.006900079
#> 296 8.75 7.25 25 0.006797976
#> 297 9.25 6.75 25 0.006474957
#> 298 9.75 6.75 25 0.006461208
#> 299 9.25 7.25 25 0.006462591
#> 300 9.75 7.25 25 0.006460035
#> 301 0.25 7.75 25 0.006427690
#> 302 0.75 7.75 25 0.006428443
#> 303 1.25 7.75 25 0.006532510
#> 304 1.75 7.75 25 0.006540773
#> 305 2.25 7.75 25 0.007046284
#> 306 2.75 7.75 25 0.008612895
#> 307 3.25 7.75 25 0.009396860
#> 308 3.75 7.75 25 0.013569572
#> 309 4.25 7.75 25 0.016157384
#> 310 4.75 7.75 25 0.019248810
#> 311 5.25 7.75 25 0.018424877
#> 312 5.75 7.75 25 0.014500459
#> 313 6.25 7.75 25 0.012355573
#> 314 6.75 7.75 25 0.008938675
#> 315 7.25 7.75 25 0.008537867
#> 316 7.75 7.75 25 0.007052420
#> 317 8.25 7.75 25 0.006544832
#> 318 8.75 7.75 25 0.006532162
#> 319 9.25 7.75 25 0.006427817
#> 320 9.75 7.75 25 0.006427475
#> 321 0.25 8.25 25 0.006414972
#> 322 0.75 8.25 25 0.006415016
#> 323 0.25 8.75 25 0.006414959
#> 324 0.75 8.75 25 0.006414961
#> 325 1.25 8.25 25 0.006436542
#> 326 1.75 8.25 25 0.006437948
#> 327 1.25 8.75 25 0.006436318
#> 328 1.75 8.75 25 0.006436588
#> 329 2.25 8.25 25 0.006542512
#> 330 2.25 8.75 25 0.006532381
#> 331 2.75 8.25 25 0.006859624
#> 332 3.25 8.25 25 0.007153462
#> 333 2.75 8.75 25 0.006798188
#> 334 3.25 8.75 25 0.006886787
#> 335 3.75 8.25 25 0.008310579
#> 336 4.25 8.25 25 0.009588580
#> 337 3.75 8.75 25 0.007510139
#> 338 4.25 8.75 25 0.007956281
#> 339 4.75 8.25 25 0.010745173
#> 340 5.25 8.25 25 0.010468382
#> 341 4.75 8.75 25 0.008513761
#> 342 5.25 8.75 25 0.008444010
#> 343 5.75 8.25 25 0.009030632
#> 344 6.25 8.25 25 0.007892865
#> 345 5.75 8.75 25 0.007816987
#> 346 6.25 8.75 25 0.007407272
#> 347 6.75 8.25 25 0.006980368
#> 348 7.25 8.25 25 0.006817889
#> 349 6.75 8.75 25 0.006844377
#> 350 7.25 8.75 25 0.006787594
#> 351 7.75 8.25 25 0.006537183
#> 352 7.75 8.75 25 0.006530765
#> 353 8.25 8.25 25 0.006437785
#> 354 8.75 8.25 25 0.006436558
#> 355 8.25 8.75 25 0.006436450
#> 356 8.75 8.75 25 0.006436316
#> 357 9.25 8.25 25 0.006415000
#> 358 9.75 8.25 25 0.006414967
#> 359 9.25 8.75 25 0.006414965
#> 360 9.75 8.75 25 0.006414962
#> 361 0.25 9.25 25 0.006412076
#> 362 0.75 9.25 25 0.006412077
#> 363 0.25 9.75 25 0.006412076
#> 364 0.75 9.75 25 0.006412076
#> 365 1.25 9.25 25 0.006414961
#> 366 1.75 9.25 25 0.006415006
#> 367 1.25 9.75 25 0.006414959
#> 368 1.75 9.75 25 0.006414964
#> 369 2.25 9.25 25 0.006427990
#> 370 2.25 9.75 25 0.006427506
#> 371 2.75 9.25 25 0.006464191
#> 372 3.25 9.25 25 0.006483190
#> 373 2.75 9.75 25 0.006460378
#> 374 3.25 9.75 25 0.006463140
#> 375 3.75 9.25 25 0.006587291
#> 376 4.25 9.25 25 0.006691139
#> 377 3.75 9.75 25 0.006517767
#> 378 4.25 9.75 25 0.006533560
#> 379 4.75 9.25 25 0.006802856
#> 380 5.25 9.25 25 0.006790097
#> 381 4.75 9.75 25 0.006570862
#> 382 5.25 9.75 25 0.006569203
#> 383 5.75 9.25 25 0.006665957
#> 384 6.25 9.25 25 0.006569098
#> 385 5.75 9.75 25 0.006530319
#> 386 6.25 9.75 25 0.006515471
#> 387 6.75 9.25 25 0.006475899
#> 388 7.25 9.25 25 0.006462416
#> 389 6.75 9.75 25 0.006462245
#> 390 7.25 9.75 25 0.006460168
#> 391 7.75 9.25 25 0.006427719
#> 392 7.75 9.75 25 0.006427475
#> 393 8.25 9.25 25 0.006414981
#> 394 8.75 9.25 25 0.006414961
#> 395 8.25 9.75 25 0.006414962
#> 396 8.75 9.75 25 0.006414959
#> 397 9.25 9.25 25 0.006412078
#> 398 9.75 9.25 25 0.006412078
#> 399 9.25 9.75 25 0.006412077
#> 400 9.75 9.75 25 0.006412077