Compute the conditional intensity at a location

Compute the conditional intensity at a location

Usage

temporal_conditional_intensity(hawkes, parameters, coordinates, step = 0.1)

Arguments

hawkes

A hawkes object.

parameters

Named list containing background, triggering, spatial, and temporal parameters.

coordinates

Numeric vector of length two giving the evaluation location.

step

Grid step size used to build the temporal evaluation grid.

Value

A numeric vector.

Examples

spatial_region <- create_rectangular_sf(0, 10, 0, 10)

params <- list(
  background_rate = list(intercept = -4),
  triggering_rate = 0.75,
  spatial = list(mean = 0, sd = 0.75),
  temporal = list(rate = 2)
)
hawkes <- rHawkes(params, time_window = c(0, 50), spatial_region = spatial_region)
temporal_conditional_intensity(hawkes, params, c(5, 5))
#>     x y          t  intensity
#> 1   5 5  0.0000000 0.01831564
#> 2   5 5  0.1000000 0.01831564
#> 3   5 5  0.1927864 0.01831567
#> 4   5 5  0.2000000 0.01831566
#> 5   5 5  0.2054320 0.17613191
#> 6   5 5  0.3000000 0.14893606
#> 7   5 5  0.4000000 0.12525860
#> 8   5 5  0.4172661 0.12458924
#> 9   5 5  0.5000000 0.10838221
#> 10  5 5  0.6000000 0.09205591
#> 11  5 5  0.6180590 0.08944008
#> 12  5 5  0.6585716 0.24022335
#> 13  5 5  0.6762628 0.23254200
#> 14  5 5  0.7000000 0.22260938
#> 15  5 5  0.7937018 0.39650398
#> 16  5 5  0.8000000 0.39177001
#> 17  5 5  0.8830540 0.38682726
#> 18  5 5  0.9000000 0.37454693
#> 19  5 5  1.0000000 0.30997315
#> 20  5 5  1.1000000 0.25710461
#> 21  5 5  1.2000000 0.21381952
#> 22  5 5  1.2706101 0.27920974
#> 23  5 5  1.2980904 0.27154838
#> 24  5 5  1.3000000 0.27058308
#> 25  5 5  1.3879274 0.23534252
#> 26  5 5  1.4000000 0.23016512
#> 27  5 5  1.4228480 0.23754834
#> 28  5 5  1.4660982 0.38803350
#> 29  5 5  1.4983513 0.36551173
#> 30  5 5  1.5000000 0.36436877
#> 31  5 5  1.6000000 0.30163998
#> 32  5 5  1.7000000 0.25028199
#> 33  5 5  1.8000000 0.20823362
#> 34  5 5  1.9000000 0.17380733
#> 35  5 5  1.9261617 0.53168830
#> 36  5 5  1.9584581 0.58369065
#> 37  5 5  2.0000000 0.53861553
#> 38  5 5  2.0037040 0.55001649
#> 39  5 5  2.1000000 0.45687230
#> 40  5 5  2.1471236 0.41742756
#> 41  5 5  2.2000000 0.37737547
#> 42  5 5  2.2758215 0.32915820
#> 43  5 5  2.3000000 0.31448442
#> 44  5 5  2.3451727 0.28890003
#> 45  5 5  2.3668537 0.28157247
#> 46  5 5  2.4000000 0.26468640
#> 47  5 5  2.5000000 0.22002695
#> 48  5 5  2.5152006 0.21398697
#> 49  5 5  2.6000000 0.18346290
#> 50  5 5  2.6092104 0.18044860
#> 51  5 5  2.6376264 0.17149144
#> 52  5 5  2.7000000 0.15352698
#> 53  5 5  2.8000000 0.12901732
#> 54  5 5  2.8487299 0.11873743
#> 55  5 5  2.9000000 0.10895051
#> 56  5 5  2.9711580 0.09692753
#> 57  5 5  2.9728445 0.09667088
#> 58  5 5  2.9784607 0.09580258
#> 59  5 5  3.0000000 0.09253542
#> 60  5 5  3.1000000 0.07908166
#> 61  5 5  3.1095671 0.07793002
#> 62  5 5  3.1586418 0.07235694
#> 63  5 5  3.1715438 0.07098171
#> 64  5 5  3.2000000 0.06806805
#> 65  5 5  3.3000000 0.05904947
#> 66  5 5  3.3132922 0.05798085
#> 67  5 5  3.3229661 0.05722219
#> 68  5 5  3.4000000 0.05166687
#> 69  5 5  3.4253624 0.05001734
#> 70  5 5  3.4457221 0.04875239
#> 71  5 5  3.5000000 0.04562133
#> 72  5 5  3.5563650 0.04272013
#> 73  5 5  3.6000000 0.04068063
#> 74  5 5  3.7000000 0.03662655
#> 75  5 5  3.7687918 0.03427289
#> 76  5 5  3.8000000 0.03330734
#> 77  5 5  3.9000000 0.03058981
#> 78  5 5  4.0000000 0.02836488
#> 79  5 5  4.1000000 0.02654326
#> 80  5 5  4.1244432 0.02615079
#> 81  5 5  4.2000000 0.02505191
#> 82  5 5  4.3000000 0.02383083
#> 83  5 5  4.3620390 0.02318727
#> 84  5 5  4.3884173 0.02293994
#> 85  5 5  4.3909790 0.02307212
#> 86  5 5  4.4000000 0.02298707
#> 87  5 5  4.4541542 0.02250765
#> 88  5 5  4.5000000 0.02214037
#> 89  5 5  4.5253472 0.02195132
#> 90  5 5  4.6000000 0.02144708
#> 91  5 5  4.7000000 0.02087944
#> 92  5 5  4.8000000 0.02041470
#> 93  5 5  4.8642307 0.03777729
#> 94  5 5  4.9000000 0.03643367
#> 95  5 5  5.0000000 0.03314943
#> 96  5 5  5.0521925 0.03194272
#> 97  5 5  5.0721443 0.04213640
#> 98  5 5  5.1000000 0.04084560
#> 99  5 5  5.1458986 0.20032854
#> 100 5 5  5.1834545 0.27470773
#> 101 5 5  5.2000000 0.26636230
#> 102 5 5  5.3000000 0.22139907
#> 103 5 5  5.4000000 0.18458629
#> 104 5 5  5.5000000 0.15444653
#> 105 5 5  5.5002579 0.28880435
#> 106 5 5  5.5606395 0.25818310
#> 107 5 5  5.6000000 0.24002462
#> 108 5 5  5.6526245 0.50647993
#> 109 5 5  5.7000000 0.46234955
#> 110 5 5  5.7195794 0.63271083
#> 111 5 5  5.8000000 0.54142842
#> 112 5 5  5.9000000 0.44660416
#> 113 5 5  6.0000000 0.36896862
#> 114 5 5  6.1000000 0.30540602
#> 115 5 5  6.2000000 0.25336536
#> 116 5 5  6.3000000 0.21075808
#> 117 5 5  6.3059536 0.21106919
#> 118 5 5  6.4000000 0.17801925
#> 119 5 5  6.5000000 0.14906990
#> 120 5 5  6.6000000 0.12536817
#> 121 5 5  6.6594116 0.23425405
#> 122 5 5  6.7000000 0.21741749
#> 123 5 5  6.7181773 0.21030921
#> 124 5 5  6.7283927 0.24130848
#> 125 5 5  6.8000000 0.21155413
#> 126 5 5  6.8269701 0.20171871
#> 127 5 5  6.9000000 0.17679535
#> 128 5 5  7.0000000 0.14806785
#> 129 5 5  7.0429830 0.22223232
#> 130 5 5  7.1000000 0.20025573
#> 131 5 5  7.1004128 0.29019405
#> 132 5 5  7.1408626 0.42073685
#> 133 5 5  7.2000000 0.37584753
#> 134 5 5  7.2444126 0.38222591
#> 135 5 5  7.2475233 0.45248394
#> 136 5 5  7.2529167 0.44782580
#> 137 5 5  7.2766092 0.42796572
#> 138 5 5  7.3000000 0.40924300
#> 139 5 5  7.4000000 0.33837989
#> 140 5 5  7.5000000 0.28036208
#> 141 5 5  7.6000000 0.23286112
#> 142 5 5  7.7000000 0.19397062
#> 143 5 5  7.8000000 0.16212978
#> 144 5 5  7.9000000 0.13606070
#> 145 5 5  8.0000000 0.11471714
#> 146 5 5  8.1000000 0.09724251
#> 147 5 5  8.2000000 0.08293550
#> 148 5 5  8.3000000 0.07122190
#> 149 5 5  8.4000000 0.06163162
#> 150 5 5  8.4343891 0.06121511
#> 151 5 5  8.4938567 0.05640459
#> 152 5 5  8.5000000 0.05593947
#> 153 5 5  8.6000000 0.04911942
#> 154 5 5  8.7000000 0.04353565
#> 155 5 5  8.8000000 0.03896403
#> 156 5 5  8.9000000 0.03522111
#> 157 5 5  9.0000000 0.03215667
#> 158 5 5  9.1000000 0.02964772
#> 159 5 5  9.1279014 0.03496015
#> 160 5 5  9.1399173 0.03456492
#> 161 5 5  9.1548038 0.03408941
#> 162 5 5  9.1578457 0.03399374
#> 163 5 5  9.2000000 0.03272612
#> 164 5 5  9.2564449 0.03264454
#> 165 5 5  9.2675944 0.03233157
#> 166 5 5  9.2853011 0.03185180
#> 167 5 5  9.3000000 0.03145966
#> 168 5 5  9.3044299 0.03134372
#> 169 5 5  9.3830150 0.02949194
#> 170 5 5  9.3949698 0.02922792
#> 171 5 5  9.4000000 0.02911869
#> 172 5 5  9.4400142 0.02828878
#> 173 5 5  9.5000000 0.02716127
#> 174 5 5  9.5553145 0.02623489
#> 175 5 5  9.6000000 0.02555785
#> 176 5 5  9.7000000 0.02424506
#> 177 5 5  9.7036894 0.02420264
#> 178 5 5  9.7488726 0.02529131
#> 179 5 5  9.8000000 0.02461327
#> 180 5 5  9.8090483 0.02450070
#> 181 5 5  9.9000000 0.02347201
#> 182 5 5  9.9164727 0.02330491
#> 183 5 5  9.9256572 0.02323908
#> 184 5 5 10.0000000 0.02255886
#> 185 5 5 10.0668811 0.02203016
#> 186 5 5 10.1000000 0.02179209
#> 187 5 5 10.1615975 0.02139276
#> 188 5 5 10.2000000 0.02116527
#> 189 5 5 10.3000000 0.02064872
#> 190 5 5 10.3260852 0.02053466
#> 191 5 5 10.3388838 0.02048488
#> 192 5 5 10.4000000 0.02023529
#> 193 5 5 10.5000000 0.01988732
#> 194 5 5 10.6000000 0.01960242
#> 195 5 5 10.6142045 0.11476866
#> 196 5 5 10.6322246 0.11135452
#> 197 5 5 10.6481712 0.10846101
#> 198 5 5 10.6520700 0.10776142
#> 199 5 5 10.7000000 0.09958529
#> 200 5 5 10.7935808 0.08587933
#> 201 5 5 10.8000000 0.08501747
#> 202 5 5 10.8544857 0.07813091
#> 203 5 5 10.8591102 0.08377886
#> 204 5 5 10.9000000 0.07863836
#> 205 5 5 10.9611979 0.07168910
#> 206 5 5 11.0000000 0.06770373
#> 207 5 5 11.0208605 0.06569168
#> 208 5 5 11.0294206 0.06504870
#> 209 5 5 11.0392307 0.06417093
#> 210 5 5 11.0508380 0.06311876
#> 211 5 5 11.0596811 0.06259521
#> 212 5 5 11.0661035 0.17949566
#> 213 5 5 11.1000000 0.16893093
#> 214 5 5 11.1409980 0.15707471
#> 215 5 5 11.1538214 0.28141226
#> 216 5 5 11.2000000 0.25820170
#> 217 5 5 11.2150691 0.25134721
#> 218 5 5 11.3000000 0.21494335
#> 219 5 5 11.3667275 0.19037804
#> 220 5 5 11.4000000 0.17930079
#> 221 5 5 11.5000000 0.15011914
#> 222 5 5 11.5667706 0.13364336
#> 223 5 5 11.6000000 0.12622796
#> 224 5 5 11.6095132 0.14630507
#> 225 5 5 11.6852554 0.12831378
#> 226 5 5 11.7000000 0.12511737
#> 227 5 5 11.8000000 0.10575750
#> 228 5 5 11.9000000 0.08990698
#> 229 5 5 12.0000000 0.07692967
#> 230 5 5 12.0495868 0.07252426
#> 231 5 5 12.1000000 0.06732510
#> 232 5 5 12.1203149 0.06537378
#> 233 5 5 12.2000000 0.05844121
#> 234 5 5 12.2394809 0.05540889
#> 235 5 5 12.3000000 0.05118027
#> 236 5 5 12.4000000 0.04522292
#> 237 5 5 12.5000000 0.04034546
#> 238 5 5 12.6000000 0.03635213
#> 239 5 5 12.6520543 0.03456881
#> 240 5 5 12.6710926 0.03396221
#> 241 5 5 12.6892573 0.03628898
#> 242 5 5 12.7000000 0.03590693
#> 243 5 5 12.7064220 0.03965113
#> 244 5 5 12.8000000 0.03600947
#> 245 5 5 12.9000000 0.03280212
#> 246 5 5 12.9923616 0.04821642
#> 247 5 5 13.0000000 0.04776310
#> 248 5 5 13.0336822 0.04589780
#> 249 5 5 13.0464874 0.04520038
#> 250 5 5 13.1000000 0.04247166
#> 251 5 5 13.2000000 0.03809292
#> 252 5 5 13.2716929 0.03545107
#> 253 5 5 13.3000000 0.03450791
#> 254 5 5 13.4000000 0.03157275
#> 255 5 5 13.4412420 0.03184012
#> 256 5 5 13.4886813 0.03061593
#> 257 5 5 13.5000000 0.03034061
#> 258 5 5 13.5782835 0.02859788
#> 259 5 5 13.6000000 0.02816085
#> 260 5 5 13.6964069 0.02643435
#> 261 5 5 13.7000000 0.02637622
#> 262 5 5 13.7305958 0.02595812
#> 263 5 5 13.8000000 0.02496761
#> 264 5 5 13.8032588 0.02586688
#> 265 5 5 13.8348235 0.02726714
#> 266 5 5 13.9000000 0.02617314
#> 267 5 5 13.9949929 0.02481358
#> 268 5 5 14.0000000 0.02474883
#> 269 5 5 14.0517628 0.02411614
#> 270 5 5 14.1000000 0.02358269
#> 271 5 5 14.2000000 0.02262793
#> 272 5 5 14.2260524 0.02240900
#> 273 5 5 14.2439725 0.02226489
#> 274 5 5 14.2973157 0.02186525
#> 275 5 5 14.3000000 0.02184625
#> 276 5 5 14.4000000 0.02120626
#> 277 5 5 14.5000000 0.02068228
#> 278 5 5 14.5081754 0.02075825
#> 279 5 5 14.5390158 0.02577304
#> 280 5 5 14.6000000 0.02491675
#> 281 5 5 14.6860034 0.02387360
#> 282 5 5 14.7000000 0.02372018
#> 283 5 5 14.7710571 0.02300420
#> 284 5 5 14.7890161 0.02283878
#> 285 5 5 14.8000000 0.02274050
#> 286 5 5 14.9000000 0.02193841
#> 287 5 5 14.9814941 0.02139355
#> 288 5 5 15.0000000 0.02128171
#> 289 5 5 15.1000000 0.02074406
#> 290 5 5 15.2000000 0.02030386
#> 291 5 5 15.2055043 0.02028209
#> 292 5 5 15.2653138 0.02006039
#> 293 5 5 15.2669135 0.02005485
#> 294 5 5 15.2828712 0.02118661
#> 295 5 5 15.3000000 0.02108992
#> 296 5 5 15.3962071 0.02060824
#> 297 5 5 15.4000000 0.02059092
#> 298 5 5 15.4321267 0.02060439
#> 299 5 5 15.5000000 0.02031386
#> 300 5 5 15.5152424 0.02064848
#> 301 5 5 15.5424789 0.02053086
#> 302 5 5 15.5531364 0.02050775
#> 303 5 5 15.6000000 0.02031162
#> 304 5 5 15.7000000 0.01994981
#> 305 5 5 15.8000000 0.01965358
#> 306 5 5 15.9000000 0.01941106
#> 307 5 5 15.9068571 0.02037297
#> 308 5 5 16.0000000 0.02002329
#> 309 5 5 16.0356889 0.02049174
#> 310 5 5 16.0840263 0.02068667
#> 311 5 5 16.1000000 0.02061212
#> 312 5 5 16.1740208 0.02030723
#> 313 5 5 16.2000000 0.02020640
#> 314 5 5 16.2051762 0.07901747
#> 315 5 5 16.3000000 0.06853127
#> 316 5 5 16.4000000 0.05942872
#> 317 5 5 16.4431905 0.06146878
#> 318 5 5 16.4983732 0.06216084
#> 319 5 5 16.5000000 0.06201841
#> 320 5 5 16.5090492 0.06123474
#> 321 5 5 16.5166793 0.06058476
#> 322 5 5 16.5499107 0.05823288
#> 323 5 5 16.6000000 0.05442780
#> 324 5 5 16.6016855 0.05523420
#> 325 5 5 16.6083822 0.05513312
#> 326 5 5 16.6423378 0.05314626
#> 327 5 5 16.7000000 0.04935240
#> 328 5 5 16.7046501 0.04906524
#> 329 5 5 16.8000000 0.04372651
#> 330 5 5 16.9000000 0.03912030
#> 331 5 5 16.9320503 0.04416317
#> 332 5 5 16.9835930 0.04167303
#> 333 5 5 17.0000000 0.04091902
#> 334 5 5 17.0572892 0.16274019
#> 335 5 5 17.1000000 0.15091544
#> 336 5 5 17.1581396 0.15224044
#> 337 5 5 17.1595504 0.15466282
#> 338 5 5 17.2000000 0.14406683
#> 339 5 5 17.2372288 0.42484340
#> 340 5 5 17.3000000 0.37688058
#> 341 5 5 17.3343382 0.36310099
#> 342 5 5 17.3531655 0.35036930
#> 343 5 5 17.3626849 0.34413779
#> 344 5 5 17.3786500 0.39115185
#> 345 5 5 17.3853396 0.60768189
#> 346 5 5 17.4000000 0.59065212
#> 347 5 5 17.4218829 0.56694050
#> 348 5 5 17.4291686 0.58395854
#> 349 5 5 17.4808735 0.54065198
#> 350 5 5 17.4850744 0.54229255
#> 351 5 5 17.5000000 0.52688241
#> 352 5 5 17.6000000 0.43469489
#> 353 5 5 17.7000000 0.35921814
#> 354 5 5 17.7421165 0.58866170
#> 355 5 5 17.7509275 0.57955674
#> 356 5 5 17.8000000 0.52709046
#> 357 5 5 17.9000000 0.43486523
#> 358 5 5 18.0000000 0.35935760
#> 359 5 5 18.0263573 0.37777068
#> 360 5 5 18.1000000 0.32854244
#> 361 5 5 18.2000000 0.27230786
#> 362 5 5 18.3000000 0.22626688
#> 363 5 5 18.4000000 0.18857172
#> 364 5 5 18.5000000 0.15770953
#> 365 5 5 18.5743765 0.13844277
#> 366 5 5 18.6000000 0.13244170
#> 367 5 5 18.7000000 0.11175416
#> 368 5 5 18.8000000 0.09481663
#> 369 5 5 18.9000000 0.08094935
#> 370 5 5 19.0000000 0.06959578
#> 371 5 5 19.0680859 0.07411960
#> 372 5 5 19.1000000 0.07066902
#> 373 5 5 19.1450504 0.43877430
#> 374 5 5 19.2000000 0.39501485
#> 375 5 5 19.3000000 0.32673087
#> 376 5 5 19.3985038 0.38511132
#> 377 5 5 19.4000000 0.38401540
#> 378 5 5 19.5000000 0.31772528
#> 379 5 5 19.5623205 0.51355395
#> 380 5 5 19.6000000 0.47760486
#> 381 5 5 19.6130297 0.56136549
#> 382 5 5 19.6407559 0.53571384
#> 383 5 5 19.7000000 0.47790093
#> 384 5 5 19.8000000 0.39459225
#> 385 5 5 19.9000000 0.32638487
#> 386 5 5 20.0000000 0.27054139
#> 387 5 5 20.1000000 0.22482062
#> 388 5 5 20.1285326 0.45422639
#> 389 5 5 20.2000000 0.39616772
#> 390 5 5 20.3000000 0.32767476
#> 391 5 5 20.4000000 0.27159746
#> 392 5 5 20.5000000 0.22568526
#> 393 5 5 20.6000000 0.18809552
#> 394 5 5 20.6001714 0.18804484
#> 395 5 5 20.7000000 0.15732581
#> 396 5 5 20.8000000 0.13212754
#> 397 5 5 20.8808716 0.18301184
#> 398 5 5 20.8923995 0.17926229
#> 399 5 5 20.8991852 0.17709279
#> 400 5 5 20.9000000 0.17683425
#> 401 5 5 20.9816271 0.15327002
#> 402 5 5 20.9924665 0.15037612
#> 403 5 5 20.9953815 0.14962263
#> 404 5 5 21.0000000 0.14841533
#> 405 5 5 21.0582250 0.13411413
#> 406 5 5 21.1000000 0.12483231
#> 407 5 5 21.2000000 0.10552411
#> 408 5 5 21.2807099 0.09281132
#> 409 5 5 21.3000000 0.08999200
#> 410 5 5 21.4000000 0.07699928
#> 411 5 5 21.4945562 0.06697795
#> 412 5 5 21.5000000 0.06645100
#> 413 5 5 21.6000000 0.05772554
#> 414 5 5 21.6933098 0.05101637
#> 415 5 5 21.7000000 0.05058174
#> 416 5 5 21.8000000 0.04473289
#> 417 5 5 21.9000000 0.03994425
#> 418 5 5 22.0000000 0.03602365
#> 419 5 5 22.1000000 0.03281373
#> 420 5 5 22.2000000 0.03018567
#> 421 5 5 22.2703596 0.02863018
#> 422 5 5 22.3000000 0.02803650
#> 423 5 5 22.4000000 0.02627441
#> 424 5 5 22.4569385 0.02542750
#> 425 5 5 22.5000000 0.02484064
#> 426 5 5 22.5846134 0.02407993
#> 427 5 5 22.6000000 0.02390525
#> 428 5 5 22.7000000 0.02289202
#> 429 5 5 22.8000000 0.02206247
#> 430 5 5 22.9000000 0.02138328
#> 431 5 5 23.0000000 0.02082721
#> 432 5 5 23.1000000 0.02037194
#> 433 5 5 23.2000000 0.01999920
#> 434 5 5 23.3000000 0.01969402
#> 435 5 5 23.3962304 0.01951478
#> 436 5 5 23.4000000 0.01950578
#> 437 5 5 23.4600963 0.01942375
#> 438 5 5 23.5000000 0.01933875
#> 439 5 5 23.5743316 0.01919744
#> 440 5 5 23.6000000 0.01915331
#> 441 5 5 23.7000000 0.01900147
#> 442 5 5 23.7673518 0.02044456
#> 443 5 5 23.7826101 0.02038069
#> 444 5 5 23.8000000 0.02031011
#> 445 5 5 23.8102898 0.02026951
#> 446 5 5 23.9000000 0.01994859
#> 447 5 5 23.9603004 0.01976307
#> 448 5 5 24.0000000 0.01965259
#> 449 5 5 24.1000000 0.01941024
#> 450 5 5 24.1303688 0.01941578
#> 451 5 5 24.1657694 0.01947264
#> 452 5 5 24.1777625 0.01944529
#> 453 5 5 24.1838950 0.01943153
#> 454 5 5 24.1939367 0.01942481
#> 455 5 5 24.2000000 0.01941144
#> 456 5 5 24.2137360 0.01938175
#> 457 5 5 24.3000000 0.01921280
#> 458 5 5 24.3813563 0.01907808
#> 459 5 5 24.4000000 0.01905018
#> 460 5 5 24.5000000 0.01891703
#> 461 5 5 24.6000000 0.01880801
#> 462 5 5 24.6612339 0.01875154
#> 463 5 5 24.6691951 0.01874467
#> 464 5 5 24.7000000 0.01871903
#> 465 5 5 24.7281853 0.01869692
#> 466 5 5 24.7907942 0.01865368
#> 467 5 5 24.8000000 0.01864752
#> 468 5 5 24.8716482 0.08094787
#> 469 5 5 24.9000000 0.07749522
#> 470 5 5 24.9698077 0.06978368
#> 471 5 5 25.0000000 0.06676778
#> 472 5 5 25.0778374 0.05978370
#> 473 5 5 25.0876502 0.05944536
#> 474 5 5 25.1000000 0.05844192
#> 475 5 5 25.2000000 0.05116826
#> 476 5 5 25.2488041 0.04811330
#> 477 5 5 25.3000000 0.04521327
#> 478 5 5 25.3584992 0.04224342
#> 479 5 5 25.4000000 0.04033756
#> 480 5 5 25.4079331 0.20696354
#> 481 5 5 25.4928324 0.17750332
#> 482 5 5 25.5000000 0.17523762
#> 483 5 5 25.5056837 0.17643216
#> 484 5 5 25.6000000 0.14925046
#> 485 5 5 25.6420549 0.15927202
#> 486 5 5 25.7000000 0.14384759
#> 487 5 5 25.8000000 0.12109251
#> 488 5 5 25.8275920 0.11768943
#> 489 5 5 25.9000000 0.10429200
#> 490 5 5 26.0000000 0.08870713
#> 491 5 5 26.1000000 0.07594731
#> 492 5 5 26.2000000 0.06550046
#> 493 5 5 26.2325680 0.06398050
#> 494 5 5 26.3000000 0.05821918
#> 495 5 5 26.3985614 0.06367337
#> 496 5 5 26.4000000 0.06354305
#> 497 5 5 26.4680107 0.05779607
#> 498 5 5 26.4968987 0.05559546
#> 499 5 5 26.5000000 0.05536495
#> 500 5 5 26.5077922 0.08240741
#> 501 5 5 26.5556297 0.07664109
#> 502 5 5 26.6000000 0.07168827
#> 503 5 5 26.7000000 0.06201345
#> 504 5 5 26.7501006 0.05867343
#> 505 5 5 26.7944627 0.05524761
#> 506 5 5 26.8000000 0.05484086
#> 507 5 5 26.8720168 0.04994134
#> 508 5 5 26.9000000 0.04821999
#> 509 5 5 26.9034804 0.05283239
#> 510 5 5 26.9181541 0.05183414
#> 511 5 5 27.0000000 0.04677297
#> 512 5 5 27.0115128 0.04612521
#> 513 5 5 27.0177893 0.04577829
#> 514 5 5 27.0223279 0.04553014
#> 515 5 5 27.1000000 0.04161453
#> 516 5 5 27.2000000 0.03739116
#> 517 5 5 27.2060989 0.04132272
#> 518 5 5 27.2529242 0.03926591
#> 519 5 5 27.3000000 0.03738342
#> 520 5 5 27.4000000 0.03392702
#> 521 5 5 27.5000000 0.03109715
#> 522 5 5 27.5990753 0.02879963
#> 523 5 5 27.6000000 0.02878026
#> 524 5 5 27.6131142 0.02850938
#> 525 5 5 27.6259020 0.02825197
#> 526 5 5 27.6715841 0.02781598
#> 527 5 5 27.7000000 0.02729111
#> 528 5 5 27.7017571 0.02725962
#> 529 5 5 27.7203086 0.02695442
#> 530 5 5 27.8000000 0.02568167
#> 531 5 5 27.8828277 0.02455945
#> 532 5 5 27.9000000 0.02434865
#> 533 5 5 28.0000000 0.02325505
#> 534 5 5 28.0405874 0.02286994
#> 535 5 5 28.0956788 0.02269346
#> 536 5 5 28.0968048 0.02268361
#> 537 5 5 28.1000000 0.02265579
#> 538 5 5 28.1465212 0.02368604
#> 539 5 5 28.2000000 0.02314129
#> 540 5 5 28.2227142 0.02292707
#> 541 5 5 28.2982400 0.02228056
#> 542 5 5 28.3000000 0.02226663
#> 543 5 5 28.3027638 0.02224485
#> 544 5 5 28.4000000 0.02155044
#> 545 5 5 28.4397262 0.02130337
#> 546 5 5 28.4499623 0.02124283
#> 547 5 5 28.4992290 0.02096869
#> 548 5 5 28.5000000 0.02096461
#> 549 5 5 28.5842577 0.02055480
#> 550 5 5 28.6000000 0.02048540
#> 551 5 5 28.7000000 0.02009209
#> 552 5 5 28.7922284 0.01979289
#> 553 5 5 28.7935663 0.01978895
#> 554 5 5 28.8000000 0.01977011
#> 555 5 5 28.9000000 0.01950646
#> 556 5 5 29.0000000 0.01929060
#> 557 5 5 29.1000000 0.01911387
#> 558 5 5 29.2000000 0.01896918
#> 559 5 5 29.2851668 0.01886682
#> 560 5 5 29.3000000 0.01885071
#> 561 5 5 29.4000000 0.01875372
#> 562 5 5 29.4305122 0.01872779
#> 563 5 5 29.5000000 0.01867431
#> 564 5 5 29.5050031 0.01867074
#> 565 5 5 29.6000000 0.01860930
#> 566 5 5 29.7000000 0.01855607
#> 567 5 5 29.8000000 0.01851248
#> 568 5 5 29.9000000 0.01847680
#> 569 5 5 30.0000000 0.01844759
#> 570 5 5 30.1000000 0.01842367
#> 571 5 5 30.2000000 0.01840409
#> 572 5 5 30.2306773 0.01839882
#> 573 5 5 30.3000000 0.01838805
#> 574 5 5 30.3104971 0.01838655
#> 575 5 5 30.4000000 0.01837493
#> 576 5 5 30.4265709 0.01837186
#> 577 5 5 30.4529202 0.01836900
#> 578 5 5 30.5000000 0.01836420
#> 579 5 5 30.5199442 0.01836230
#> 580 5 5 30.6000000 0.01835540
#> 581 5 5 30.6486235 0.01835171
#> 582 5 5 30.7000000 0.01834819
#> 583 5 5 30.7033259 0.01834798
#> 584 5 5 30.8000000 0.01834229
#> 585 5 5 30.9000000 0.01833746
#> 586 5 5 31.0000000 0.01833350
#> 587 5 5 31.0894077 0.01833058
#> 588 5 5 31.1000000 0.01833027
#> 589 5 5 31.1348663 0.01832928
#> 590 5 5 31.2000000 0.01832761
#> 591 5 5 31.3000000 0.01832544
#> 592 5 5 31.4000000 0.01832367
#> 593 5 5 31.4094203 0.01832352
#> 594 5 5 31.5000000 0.01832221
#> 595 5 5 31.6000000 0.01832102
#> 596 5 5 31.6565339 0.02816595
#> 597 5 5 31.7000000 0.02734580
#> 598 5 5 31.7043696 0.02726723
#> 599 5 5 31.7629678 0.02627996
#> 600 5 5 31.7791091 0.02621989
#> 601 5 5 31.8000000 0.02589644
#> 602 5 5 31.9000000 0.02452227
#> 603 5 5 31.9969020 0.02343362
#> 604 5 5 32.0000000 0.02340201
#> 605 5 5 32.1000000 0.02248001
#> 606 5 5 32.1401401 0.02215896
#> 607 5 5 32.2000000 0.02172532
#> 608 5 5 32.3000000 0.02110725
#> 609 5 5 32.4000000 0.02060121
#> 610 5 5 32.5000000 0.02018691
#> 611 5 5 32.6000000 0.01984771
#> 612 5 5 32.7000000 0.01956999
#> 613 5 5 32.8000000 0.01934261
#> 614 5 5 32.8571401 0.01923173
#> 615 5 5 32.9000000 0.01915648
#> 616 5 5 33.0000000 0.01900406
#> 617 5 5 33.1000000 0.01887927
#> 618 5 5 33.1213322 0.01885573
#> 619 5 5 33.1286824 0.01884785
#> 620 5 5 33.2000000 0.01877710
#> 621 5 5 33.3000000 0.01869345
#> 622 5 5 33.3239693 0.01868153
#> 623 5 5 33.4000000 0.01862991
#> 624 5 5 33.5000000 0.01857294
#> 625 5 5 33.6000000 0.01852630
#> 626 5 5 33.6737808 0.01849742
#> 627 5 5 33.7000000 0.01848813
#> 628 5 5 33.8000000 0.01845686
#> 629 5 5 33.9000000 0.01843126
#> 630 5 5 34.0000000 0.01841030
#> 631 5 5 34.0529145 0.01840080
#> 632 5 5 34.1000000 0.01839314
#> 633 5 5 34.2000000 0.01837909
#> 634 5 5 34.3000000 0.01836759
#> 635 5 5 34.3445873 0.01842714
#> 636 5 5 34.4000000 0.01841544
#> 637 5 5 34.4473829 0.01954085
#> 638 5 5 34.5000000 0.01941846
#> 639 5 5 34.6000000 0.01921856
#> 640 5 5 34.7000000 0.01905489
#> 641 5 5 34.8000000 0.01892088
#> 642 5 5 34.9000000 0.01881117
#> 643 5 5 34.9411721 0.01887373
#> 644 5 5 35.0000000 0.01881178
#> 645 5 5 35.1000000 0.01872185
#> 646 5 5 35.2000000 0.01864821
#> 647 5 5 35.2683325 0.01932884
#> 648 5 5 35.3000000 0.01926666
#> 649 5 5 35.4000000 0.01909427
#> 650 5 5 35.5000000 0.01895312
#> 651 5 5 35.6000000 0.01883757
#> 652 5 5 35.7000000 0.01874296
#> 653 5 5 35.8000000 0.01866550
#> 654 5 5 35.9000000 0.01860208
#> 655 5 5 36.0000000 0.01855016
#> 656 5 5 36.1000000 0.01850765
#> 657 5 5 36.1874128 0.01874731
#> 658 5 5 36.2000000 0.01873658
#> 659 5 5 36.2245975 0.01871646
#> 660 5 5 36.3000000 0.01866035
#> 661 5 5 36.3887373 0.01860432
#> 662 5 5 36.4000000 0.01859789
#> 663 5 5 36.5000000 0.01854673
#> 664 5 5 36.6000000 0.01850484
#> 665 5 5 36.6971126 0.01848479
#> 666 5 5 36.7000000 0.01848382
#> 667 5 5 36.7439460 0.02217884
#> 668 5 5 36.7949680 0.02201930
#> 669 5 5 36.8000000 0.02198221
#> 670 5 5 36.8355105 0.02173084
#> 671 5 5 36.8882409 0.02177808
#> 672 5 5 36.9000000 0.02169760
#> 673 5 5 37.0000000 0.02108455
#> 674 5 5 37.1000000 0.02058263
#> 675 5 5 37.1649435 0.02030867
#> 676 5 5 37.2000000 0.02017372
#> 677 5 5 37.2605805 0.01996170
#> 678 5 5 37.2801521 0.01992336
#> 679 5 5 37.2928300 0.01992704
#> 680 5 5 37.3000000 0.01990410
#> 681 5 5 37.4000000 0.01961616
#> 682 5 5 37.4383385 0.02087907
#> 683 5 5 37.5000000 0.02058165
#> 684 5 5 37.5103050 0.02053543
#> 685 5 5 37.6000000 0.02017089
#> 686 5 5 37.7000000 0.01983459
#> 687 5 5 37.8000000 0.01955925
#> 688 5 5 37.8284560 0.03270043
#> 689 5 5 37.9000000 0.03078262
#> 690 5 5 37.9781741 0.09788809
#> 691 5 5 38.0000000 0.09448933
#> 692 5 5 38.0450761 0.08792256
#> 693 5 5 38.1000000 0.08068139
#> 694 5 5 38.2000000 0.06937639
#> 695 5 5 38.3000000 0.06012065
#> 696 5 5 38.4000000 0.05254269
#> 697 5 5 38.5000000 0.04633838
#> 698 5 5 38.6000000 0.04125871
#> 699 5 5 38.7000000 0.03709984
#> 700 5 5 38.7216034 0.03726551
#> 701 5 5 38.7750374 0.03541478
#> 702 5 5 38.8000000 0.03458206
#> 703 5 5 38.9000000 0.03163346
#> 704 5 5 39.0000000 0.02921935
#> 705 5 5 39.1000000 0.02724284
#> 706 5 5 39.2000000 0.02562461
#> 707 5 5 39.2593690 0.02480630
#> 708 5 5 39.3000000 0.02429972
#> 709 5 5 39.4000000 0.02321499
#> 710 5 5 39.4965608 0.02454346
#> 711 5 5 39.5000000 0.02450077
#> 712 5 5 39.5921556 0.02345967
#> 713 5 5 39.6000000 0.02337960
#> 714 5 5 39.6208376 0.02336511
#> 715 5 5 39.6301482 0.02327195
#> 716 5 5 39.6360997 0.02442690
#> 717 5 5 39.6384309 0.02439848
#> 718 5 5 39.7000000 0.02369373
#> 719 5 5 39.7561659 0.02312230
#> 720 5 5 39.8000000 0.02271885
#> 721 5 5 39.9000000 0.02192068
#> 722 5 5 40.0000000 0.02126720
#> 723 5 5 40.1000000 0.02073217
#> 724 5 5 40.2000000 0.02029413
#> 725 5 5 40.3000000 0.01993549
#> 726 5 5 40.4000000 0.01964186
#> 727 5 5 40.5000000 0.01940146
#> 728 5 5 40.6000000 0.01920463
#> 729 5 5 40.7000000 0.01904348
#> 730 5 5 40.8000000 0.01891155
#> 731 5 5 40.9000000 0.01880353
#> 732 5 5 40.9927317 0.01872524
#> 733 5 5 41.0000000 0.01871933
#> 734 5 5 41.1000000 0.01864615
#> 735 5 5 41.2000000 0.01858624
#> 736 5 5 41.3000000 0.01853719
#> 737 5 5 41.3088201 0.01853332
#> 738 5 5 41.4000000 0.01849703
#> 739 5 5 41.5000000 0.01846415
#> 740 5 5 41.6000000 0.01843723
#> 741 5 5 41.7000000 0.01841519
#> 742 5 5 41.8000000 0.01839714
#> 743 5 5 41.9000000 0.01838237
#> 744 5 5 42.0000000 0.01837027
#> 745 5 5 42.1000000 0.01836037
#> 746 5 5 42.2000000 0.01835226
#> 747 5 5 42.3000000 0.01834562
#> 748 5 5 42.4000000 0.01834019
#> 749 5 5 42.5000000 0.01833574
#> 750 5 5 42.6000000 0.01833209
#> 751 5 5 42.7000000 0.01832911
#> 752 5 5 42.8000000 0.01832667
#> 753 5 5 42.9000000 0.01832467
#> 754 5 5 43.0000000 0.01832303
#> 755 5 5 43.0010439 0.02791913
#> 756 5 5 43.1000000 0.02619474
#> 757 5 5 43.2000000 0.02476650
#> 758 5 5 43.3000000 0.02359716
#> 759 5 5 43.4000000 0.02263978
#> 760 5 5 43.5000000 0.02185595
#> 761 5 5 43.6000000 0.02121420
#> 762 5 5 43.7000000 0.02068878
#> 763 5 5 43.8000000 0.02025860
#> 764 5 5 43.9000000 0.01990640
#> 765 5 5 44.0000000 0.01961805
#> 766 5 5 44.1000000 0.01938196
#> 767 5 5 44.1485330 0.01928408
#> 768 5 5 44.1593778 0.01953839
#> 769 5 5 44.2000000 0.01944298
#> 770 5 5 44.3000000 0.01923862
#> 771 5 5 44.4000000 0.01907132
#> 772 5 5 44.5000000 0.01893433
#> 773 5 5 44.6000000 0.01882218
#> 774 5 5 44.7000000 0.01873036
#> 775 5 5 44.8000000 0.01865519
#> 776 5 5 44.9000000 0.01859364
#> 777 5 5 45.0000000 0.01854324
#> 778 5 5 45.0481920 0.14231842
#> 779 5 5 45.1000000 0.13011300
#> 780 5 5 45.1451911 0.13909405
#> 781 5 5 45.1845599 0.13019120
#> 782 5 5 45.2000000 0.12678926
#> 783 5 5 45.3000000 0.10712633
#> 784 5 5 45.4000000 0.09102768
#> 785 5 5 45.5000000 0.07784722
#> 786 5 5 45.6000000 0.06705598
#> 787 5 5 45.7000000 0.05822085
#> 788 5 5 45.7191034 0.05718958
#> 789 5 5 45.8000000 0.05138248
#> 790 5 5 45.8930160 0.04576929
#> 791 5 5 45.9000000 0.04538848
#> 792 5 5 46.0000000 0.04048101
#> 793 5 5 46.0998777 0.03646755
#> 794 5 5 46.1000000 0.03646311
#> 795 5 5 46.1078419 0.03618070
#> 796 5 5 46.1101708 0.03628375
#> 797 5 5 46.1393692 0.03526452
#> 798 5 5 46.2000000 0.03332900
#> 799 5 5 46.3000000 0.03060754
#> 800 5 5 46.3666995 0.02907598
#> 801 5 5 46.4000000 0.02838268
#> 802 5 5 46.4410056 0.02759001
#> 803 5 5 46.4521241 0.03000267
#> 804 5 5 46.5000000 0.02893552
#> 805 5 5 46.6000000 0.02701046
#> 806 5 5 46.7000000 0.02543436
#> 807 5 5 46.8000000 0.02414395
#> 808 5 5 46.9000000 0.02308746
#> 809 5 5 46.9041206 0.25638497
#> 810 5 5 47.0000000 0.21484329
#> 811 5 5 47.0436585 0.19844806
#> 812 5 5 47.1000000 0.17925202
#> 813 5 5 47.1735717 0.15723121
#> 814 5 5 47.2000000 0.15007929
#> 815 5 5 47.2680892 0.13330405
#> 816 5 5 47.3000000 0.12619459
#> 817 5 5 47.4000000 0.10663946
#> 818 5 5 47.5000000 0.09062906
#> 819 5 5 47.6000000 0.07752086
#> 820 5 5 47.7000000 0.06678878
#> 821 5 5 47.7541758 0.06182722
#> 822 5 5 47.8000000 0.05801673
#> 823 5 5 47.8485630 0.05434471
#> 824 5 5 47.9000000 0.05082253
#> 825 5 5 48.0000000 0.04493003
#> 826 5 5 48.0756717 0.04119211
#> 827 5 5 48.0797577 0.04100592
#> 828 5 5 48.0991597 0.04014231
#> 829 5 5 48.1000000 0.04010566
#> 830 5 5 48.1950355 0.03634261
#> 831 5 5 48.2000000 0.03616451
#> 832 5 5 48.2057850 0.03595919
#> 833 5 5 48.2728191 0.03374546
#> 834 5 5 48.3000000 0.03292906
#> 835 5 5 48.3242253 0.03223794
#> 836 5 5 48.3644171 0.03130284
#> 837 5 5 48.4000000 0.03041071
#> 838 5 5 48.4823998 0.02857307
#> 839 5 5 48.5000000 0.02821829
#> 840 5 5 48.5284911 0.02766987
#> 841 5 5 48.5460370 0.02734731
#> 842 5 5 48.5510984 0.02738026
#> 843 5 5 48.6000000 0.02653568
#> 844 5 5 48.7000000 0.02504564
#> 845 5 5 48.7923570 0.02391057
#> 846 5 5 48.8000000 0.02382570
#> 847 5 5 48.9000000 0.02282690
#> 848 5 5 49.0000000 0.02200914
#> 849 5 5 49.0615358 0.02207069
#> 850 5 5 49.1000000 0.02179266
#> 851 5 5 49.2000000 0.02116238
#> 852 5 5 49.2004936 0.02115962
#> 853 5 5 49.2769654 0.02079044
#> 854 5 5 49.3000000 0.02067901
#> 855 5 5 49.3714770 0.02036420
#> 856 5 5 49.3753900 0.02079804
#> 857 5 5 49.4000000 0.02067881
#> 858 5 5 49.4501614 0.02050688
#> 859 5 5 49.5000000 0.02029900
#> 860 5 5 49.5926926 0.01996346
#> 861 5 5 49.6000000 0.01993955
#> 862 5 5 49.6322287 0.01983818
#> 863 5 5 49.7000000 0.01964518
#> 864 5 5 49.8000000 0.01940418
#> 865 5 5 49.8770980 0.01941931
#> 866 5 5 49.9000000 0.01936990
#> 867 5 5 49.9678116 0.02393881
#> 868 5 5 50.0000000 0.02358822